% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-abundance.R
\name{calc_abundance}
\alias{calc_abundance}
\title{Calculate clonotype abundance}
\usage{
calc_abundance(
  input,
  cluster_col = NULL,
  clonotype_col = "clonotype_id",
  prefix = "",
  return_df = FALSE
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating clonotype abundance}

\item{prefix}{Prefix to add to new columns}

\item{return_df}{Return results as a data.frame. If set to FALSE, results
will be added to the input object.}
}
\value{
Single cell object or data.frame with clonotype abundance metrics
}
\description{
Calculate clonotype abundance
}
\examples{
# Calculate clonotype abundance using all cells
res <- calc_abundance(
  vdj_so,
  clonotype_col = "clonotype_id"
)

head(res@meta.data, 1)

# Group cells based on meta.data column before calculating abundance
res <- calc_abundance(
  vdj_sce,
  cluster_col = "orig.ident"
)

head(res@colData, 1)

# Add a prefix to the new columns
# this is useful if multiple abundance calculations are stored in the
# meta.data
res <- calc_abundance(
  vdj_so,
  prefix = "bcr_"
)

head(res@meta.data, 1)

# Return a data.frame instead of adding the results to the input object
res <- calc_abundance(
  vdj_sce,
  return_df = TRUE
)

head(res, 1)

}
