% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diversity.R
\name{calc_diversity}
\alias{calc_diversity}
\alias{calc_diversity.default}
\alias{calc_diversity.Seurat}
\title{Calculate repertoire diversity}
\usage{
calc_diversity(input, ...)

\method{calc_diversity}{default}(
  input,
  clonotype_col = "cdr3_nt",
  cluster_col = NULL,
  method = abdiv::simpson,
  prefix = "",
  ...
)

\method{calc_diversity}{Seurat}(
  input,
  clonotype_col = "cdr3_nt",
  cluster_col = NULL,
  method = abdiv::simpson,
  prefix = "",
  return_seurat = TRUE
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating diversity}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating diversity. If cluster_col is omitted, diversity index will be
calculated for all clonotypes.}

\item{method}{Method to use for calculating diversity. A named list can also
be passed to use multiple methods. The names should specify names for the
output columns.
names.}

\item{prefix}{Prefix to add to new meta.data columns}

\item{return_seurat}{Return a Seurat object. If set to FALSE, the meta.data
table is returned.}
}
\value{
Seurat object with diversity index added to meta.data
}
\description{
Calculate repertoire diversity
}
