% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-gene-usage.R
\name{calc_gene_pairs}
\alias{calc_gene_pairs}
\title{Calculate paired usage of V(D)J segments across chains}
\usage{
calc_gene_pairs(
  input,
  data_col,
  chains,
  cluster_col = NULL,
  chain_col = global$chain_col,
  sep = global$sep
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing V(D)J genes identified for each
clonotype.}

\item{chains}{Chains to use for calculating usage of different gene pairs.
This should be a character vector containing the two chains to use for
calculations, e.g. \code{c("IGH", "IGK")}.}

\item{cluster_col}{meta.data column containing cell clusters to use when
calculating gene usage}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}
}
\value{
data.frame containing gene pair summary
}
\description{
Quantify the paired usage of V(D)J segments across two chains. For example,
calc_gene_pairs() can calculate the frequency that different TRA and TRB V
segments appear together.
}
\examples{
# Calculate the frequency of different V genes for IGH and IGK chains
calc_gene_pairs(
  vdj_sce,
  data_col    = "v_gene",
  chains      = c("IGH", "IGK"),
  cluster_col = "orig.ident"
)

}
\seealso{
\code{\link[=plot_gene_pairs]{plot_gene_pairs()}}, \code{\link[=calc_gene_usage]{calc_gene_usage()}}, \code{\link[=plot_gene_usage]{plot_gene_usage()}}
}
