% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-gene-usage.R
\name{calc_gene_usage}
\alias{calc_gene_usage}
\title{Calculate V(D)J segment usage}
\usage{
calc_gene_usage(
  input,
  data_cols,
  cluster_col = NULL,
  chain = NULL,
  chain_col = "chains",
  sep = ";"
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{data_cols}{meta.data column containing V(D)J genes identified for each
clonotype. If multiple columns are provided, paired usage of genes will be
calculated.}

\item{cluster_col}{meta.data column containing cell clusters to use when
calculating gene usage}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}
}
\value{
data.frame containing gene usage summary
}
\description{
Calculate V(D)J segment usage
}
\examples{
# Calculate V(D)J segment usage for all cells
calc_gene_usage(
  vdj_so,
  data_cols = "v_gene"
)

# Calculate gene usage separately for cell clusters
calc_gene_usage(
  vdj_sce,
  data_cols = "v_gene",
  cluster_col = "orig.ident"
)

# Calculate gene usage for a specific chain(s)
calc_gene_usage(
  vdj_so,
  data_cols = "v_gene",
  chain = c("IGK", "IGL")
)

# Calculate paired usage of V(D)J segments
calc_gene_usage(
  vdj_sce,
  data_cols = c("v_gene", "j_gene"),
)

}
\seealso{
\code{\link[=plot_gene_usage]{plot_gene_usage()}}
}
