% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_jaccard}
\alias{calc_jaccard}
\title{Calculate Jaccard index for cell identities}
\usage{
calc_jaccard(
  sobj_in,
  clonotype_col = "clonotype_id",
  cluster_col,
  ref_cluster = NULL,
  prefix = ""
)
}
\arguments{
\item{sobj_in}{Seurat object}

\item{clonotype_col}{meta.data column containing clonotype ids}

\item{cluster_col}{meta.data column containing cell clusters}

\item{ref_cluster}{Cluster id to use as a reference for calculating Jaccard
index. If ref_cluster is omitted, Jaccard index will be calculated for all
combinations of clusters.}

\item{prefix}{Prefix to add to new meta.data columns}
}
\value{
Seurat object with Jaccard index added to meta.data
}
\description{
Calculate Jaccard index for cell identities
}
