% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{calc_mds}
\alias{calc_mds}
\title{Perform multidimensional scaling}
\usage{
calc_mds(
  input,
  data_col,
  cluster_col,
  method = abdiv::jaccard,
  chain = NULL,
  chain_col = "chains",
  prefix = "",
  return_df = FALSE,
  sep = ";"
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating repertoire overlap}

\item{method}{Method to use for comparing clusters, possible values are:
\itemize{
\item A function that takes two numeric vectors containing counts for each
unique value in the column provided to the data_col column, e.g.
abdiv::jaccard()
}}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{prefix}{Prefix to add to new columns}

\item{return_df}{Return results as a data.frame. If set to FALSE, results
will be added to the input object.}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}
}
\value{
Single cell object or data.frame with MDS coordinates
}
\description{
Perform multidimensional scaling
}
\seealso{
\code{\link[=plot_mds]{plot_mds()}}, \code{\link[MASS:isoMDS]{MASS::isoMDS()}}
}
