% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_overlap}
\alias{calc_overlap}
\title{Calculate repertoire overlap between clusters}
\usage{
calc_overlap(
  sobj_in,
  clonotype_col = "clonotype_id",
  cluster_col,
  prefix = "jcrd_",
  return_seurat = TRUE
)
}
\arguments{
\item{sobj_in}{Seurat object}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{prefix}{Prefix to add to new meta.data columns}

\item{return_seurat}{Return a Seurat object. If set to FALSE, a matrix is
returned}
}
\value{
Seurat object with Jaccard index added to meta.data
}
\description{
Calculate repertoire overlap between clusters
}
