% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-usage.R
\name{calc_usage}
\alias{calc_usage}
\title{Calculate gene usage}
\usage{
calc_usage(
  input,
  gene_cols,
  cluster_col = NULL,
  chain = NULL,
  chain_col = "chains",
  sep = ";"
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{gene_cols}{meta.data column containing genes identified for each
clonotype}

\item{cluster_col}{meta.data column containing cell clusters to use when
calculating gene usage}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}
}
\value{
data.frame containing gene usage summary
}
\description{
Calculate gene usage
}
