% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-cdr3.R
\name{cluster_seqs}
\alias{cluster_seqs}
\title{Cluster cells based on CDR3 sequences}
\usage{
cluster_seqs(
  input,
  data_col = "cdr3",
  chain,
  method = "louvain",
  k = 10,
  resolution = 0.5,
  chain_col = "chains",
  prefix = paste0(data_col, "_"),
  return_df = FALSE,
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing CDR3 sequences to use for
calculating Levenshtein distance.}

\item{chain}{Chain to use for clustering CDR3 sequences. Cells with more
than one of the provided chain will be excluded from the analysis.}

\item{method}{Method to use for clustering, possible values are:
\itemize{
\item 'louvain', multi-level optimization of modality implemented with
igraph::cluster_louvain()
\item 'leiden', the Leiden clustering algorithm implemented with
igraph::cluster_leiden()
}}

\item{k}{Number of neighbors for k-nearest neighbors algorithm}

\item{resolution}{Resolution (coarseness) of clusters}

\item{chain_col}{meta.data column containing chains for each cell.}

\item{prefix}{Prefix to add to graph name}

\item{return_df}{Return results as a data.frame. If FALSE, results will be
added to the input object.}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{...}{Additional parameters to pass to clustering method}
}
\value{
Single cell object or data.frame with clustering results
}
\description{
Cluster cells based on CDR3 sequences
}
