% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-seqs.R
\name{cluster_sequences}
\alias{cluster_sequences}
\title{Cluster cells based on CDR3 sequences}
\usage{
cluster_sequences(
  input,
  data_col = "cdr3",
  chain = NULL,
  method = "louvain",
  resolution = 0.5,
  k = 10,
  dist_method = NULL,
  run_umap = TRUE,
  chain_col = "chains",
  prefix = paste0(data_col, "_"),
  return_df = FALSE,
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing sequences to use for calculating
Levenshtein distance.}

\item{chain}{Chain to use for clustering sequences. Cells with more than one
of the provided chain will be excluded from the analysis.}

\item{method}{Method to use for clustering, possible values are:
\itemize{
\item 'louvain', multi-level optimization of modality implemented with
\code{\link[igraph:cluster_louvain]{igraph::cluster_louvain()}}
\item 'leiden', the Leiden clustering algorithm implemented with
\code{\link[igraph:cluster_leiden]{igraph::cluster_leiden()}}
}}

\item{resolution}{Resolution (coarseness) of clusters}

\item{k}{Number of neighbors for k-nearest neighbors algorithm}

\item{dist_method}{Method to use for computing distance between sequences.
If NULL, distance is calculated for amino acid sequences using the BLOSUM62
substitution matrix and levenshtein distance is calculated for nucleotide
sequences. Other possible values include:
\itemize{
\item 'levenshtein'
\item 'hamming'
\item The name of a substitution matrix available from the Biostrings package,
e.g. 'BLOSUM62'
}}

\item{run_umap}{Should the Uniform Manifold Approximation and Projection
(UMAP) dimensional reduction method be performed. This will add UMAP
coordinates to the meta.data.}

\item{chain_col}{meta.data column containing chains for each cell.}

\item{prefix}{Prefix to add to graph name}

\item{return_df}{Return results as a data.frame. If FALSE, results will be
added to the input object.}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{...}{Additional parameters to pass to clustering method}
}
\value{
Single cell object or data.frame with clustering results
}
\description{
Cluster cells based on CDR3 sequences
}
\seealso{
\code{\link[=plot_motifs]{plot_motifs()}}
}
