% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_vdj}
\alias{cluster_vdj}
\title{Cluster cells based on receptor sequence}
\usage{
cluster_vdj(sobj_in, cdr3_col = "cdr3s_aa", resolution = 0.1, ...)
}
\arguments{
\item{sobj_in}{Seurat object}

\item{cdr3_col}{meta.data column containing CDR3 sequences to use for
calculating Levenshtein distance}

\item{resolution}{Clustering resolution to pass to FindClusters}

\item{...}{Additional parameters to pass to FindClusters}
}
\value{
Seurat object with an added shared nearest neighbor graph (vdj_snn)
and a meta.data column containing cluster ids
}
\description{
Cluster cells based on receptor sequence
}
