% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_char}
\alias{convert_char}
\title{Attempt to convert character vector using provided function}
\usage{
convert_char(x, fun)
}
\arguments{
\item{x}{Character vector to convert}

\item{fun}{Function to try}
}
\value{
Value converted using fun
}
\description{
Attempt to convert character vector using provided function
}
