% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vdj.R
\name{define_clonotypes}
\alias{define_clonotypes}
\title{Define clonotypes based on V(D)J data}
\usage{
define_clonotypes(input, vdj_cols, clonotype_col = "clonotype_id")
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{vdj_cols}{meta.data columns containing V(D)J data to use for defining
clonotypes}

\item{clonotype_col}{Name of column to use for storing clonotype IDs}
}
\value{
Single cell object or data.frame with new clonotype IDs
}
\description{
Define clonotypes based on V(D)J data
}
