% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{djvdj_theme}
\alias{djvdj_theme}
\title{Theme for djvdj plotting functions}
\usage{
djvdj_theme(
  ttl_size = 12,
  txt_size = 8,
  ln_size = 0.5,
  txt_col = "black",
  ln_col = "grey85"
)
}
\arguments{
\item{ttl_size}{Size of axis titles}

\item{txt_size}{Size of axis text}

\item{ln_size}{Size of axis lines}

\item{txt_col}{Color of axis text}

\item{ln_col}{Color of axis lines}
}
\value{
ggplot theme
}
\description{
Theme for djvdj plotting functions
}
