% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.add_aes}
\alias{.add_aes}
\title{Add list of aes params to ggplot object}
\usage{
.add_aes(gg_in, add_aes, lay)
}
\arguments{
\item{gg_in}{ggplot object}

\item{add_aes}{List of aes params to add or override}

\item{lay}{Layer number to modify}
}
\value{
ggplot object
}
\description{
Add list of aes params to ggplot object
}
