% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_meta}
\alias{.add_meta}
\alias{.add_meta.default}
\alias{.add_meta.Seurat}
\alias{.add_meta.SingleCellExperiment}
\title{Add meta.data to single cell object}
\usage{
.add_meta(input, ...)

\method{.add_meta}{default}(input, meta)

\method{.add_meta}{Seurat}(input, meta)

\method{.add_meta}{SingleCellExperiment}(input, meta)
}
\arguments{
\item{input}{Object containing single cell data}

\item{...}{Arguments passed to other methods}

\item{meta}{meta.data to add to object}
}
\value{
Object with added meta.data
}
\description{
Add meta.data to single cell object
}
