% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.calc_abund}
\alias{.calc_abund}
\title{Calculate clonotype abundance}
\usage{
.calc_abund(df_in, cell_col, clone_col, clust_col = NULL)
}
\arguments{
\item{df_in}{Input data.frame}

\item{cell_col}{Column containing cell IDs}

\item{clone_col}{Column containing clonotype IDs}

\item{clust_col}{Column containing cluster IDs}
}
\value{
data.frame
}
\description{
Calculate clonotype abundance
}
