% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.call_abdiv}
\alias{.call_abdiv}
\title{Create expression for calling abdiv functions}
\usage{
.call_abdiv(ins, fn)
}
\arguments{
\item{ins}{Names of input columns or vectors}

\item{fn}{Function to use}
}
\value{
Expression to call abdiv function
}
\description{
Create expression for calling abdiv functions
}
