% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-vdj.R
\name{.check_overlap}
\alias{.check_overlap}
\title{Check cell barcode overlap with object}
\usage{
.check_overlap(input, meta, nm, pct_min = 25)
}
\arguments{
\item{input}{Single cell object}

\item{meta}{meta.data to check against object}

\item{nm}{Sample name to use for messages}

\item{pct_min}{Warn user if the percent overlap is less than pct_min}
}
\value{
input data
}
\description{
Check cell barcode overlap with object
}
