% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-vdj.R
\name{.classify_vdj}
\alias{.classify_vdj}
\title{Identify insertions and deletions for each contig}
\usage{
.classify_vdj(df_in, chain_col = "chains")
}
\arguments{
\item{df_in}{data.frame containing V(D)J data formatted so that each row
represents a single contig}

\item{chain_col}{Column in input data containing chain identity}

\item{vdj_dir}{Directory containing the output from cellranger vdj. A vector
or named vector can be given to load data from several runs. If a named
vector is given, the cell barcodes will be prefixed with the provided names.
This mimics the behavior of the Read10X function found in the Seurat
package.
Determine whether TCR or BCR data were provided}
}
\value{
Character string indicating whether TCR or BCR data were provided
}
\description{
Identify insertions and deletions for each contig
}
