% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.create_heatmap}
\alias{.create_heatmap}
\title{Create ggplot heatmap}
\usage{
.create_heatmap(
  df_in,
  x,
  y,
  fill,
  plot_colors = NULL,
  na_color = "white",
  legd_ttl = fill,
  ang = 45,
  hjst = 1,
  ...
)
}
\arguments{
\item{df_in}{data.frame}

\item{x}{Variable to plot on the x-axis}

\item{y}{Variable to plot on the y-axis}

\item{fill}{Variable to use for the fill color}

\item{plot_colors}{Character vector containing colors for plotting}

\item{na_color}{Color to use for missing values}

\item{legd_ttl}{Legend title}

\item{ang}{Angle of x-axis text}

\item{hjst}{Horizontal justification for x-axis text}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Create ggplot heatmap
}
