% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-vdj.R
\name{.extract_isotypes}
\alias{.extract_isotypes}
\title{Add isotypes to V(D)J data}
\usage{
.extract_isotypes(df_in, iso_pat = "^IGH[ADEGM]", iso_col = "c_gene")
}
\arguments{
\item{df_in}{data.frame containing V(D)J data formatted so each row
represents a single contig}

\item{iso_pat}{Regular expression to use for extracting isotypes from
iso_col}

\item{iso_col}{Column containing data to use for extracting isotypes}
}
\value{
Input data.frame with isotype column added
}
\description{
Add isotypes to V(D)J data
}
