% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vdj.R
\name{.format_cell_prefixes}
\alias{.format_cell_prefixes}
\title{Format and add cell prefixes}
\usage{
.format_cell_prefixes(vdj_dir, cell_prefix, sep = "_")
}
\arguments{
\item{vdj_dir}{Directory containing the output from cellranger vdj. A vector
or named vector can be given to load data from several runs. If a named
vector is given, the cell barcodes will be prefixed with the provided names.
This mimics the behavior of the Read10X function found in the Seurat
package. Cell barcode prefixes can also be provided using the cell_prefix
argument.}

\item{cell_prefix}{Prefix to add to cell barcodes, this is helpful when
loading data from multiple runs into a single object. If set to NULL, cell
barcode prefixes will be automatically generated in a similar way as the
Read10X function found in the Seurat package.}

\item{sep}{Separator to use when appending prefixes to cell barcodes, set to
NULL to not add a separator}
}
\value{
Paths provided to vdj_dir with cell prefixes added as names
}
\description{
Format and add cell prefixes
}
