% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_meta}
\alias{.get_meta}
\alias{.get_meta.default}
\alias{.get_meta.Seurat}
\alias{.get_meta.SingleCellExperiment}
\title{Add meta.data to single cell object}
\usage{
.get_meta(input)

\method{.get_meta}{default}(input)

\method{.get_meta}{Seurat}(input)

\method{.get_meta}{SingleCellExperiment}(input)
}
\arguments{
\item{input}{Object containing single cell data}
}
\value{
tibble containing meta.data pulled from object
}
\description{
Add meta.data to single cell object
}
