% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_vdj_cols}
\alias{.get_vdj_cols}
\title{Identify columns with V(D)J data}
\usage{
.get_vdj_cols(df_in, clone_col, cols_in, sep)
}
\arguments{
\item{df_in}{data.frame}

\item{clone_col}{Column containing clonotype IDs to use for identifying
columns with V(D)J data. If both clone_col and cols_in are set to NULL all
columns are used.}

\item{cols_in}{meta.data columns containing V(D)J data to use for filtering.
If set to NULL (the default) columns are automatically selected by
identifying columns that have NAs in the same rows as clone_col.}

\item{sep}{Separator used for storing per cell V(D)J data}
}
\value{
List with two vectors, one containing columns with V(D)J data and
the other containing columns where separator has been detected.
}
\description{
Identify columns with V(D)J data
}
