% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-vdj.R
\name{.get_vdj_path}
\alias{.get_vdj_path}
\title{Check for V(D)J data file in provided directory}
\usage{
.get_vdj_path(vdj_dir, file)
}
\arguments{
\item{vdj_dir}{Directory containing the output from cellranger vdj}

\item{file}{Name of cellranger vdj output file}
}
\value{
path to cellranger vdj output file
}
\description{
Check for V(D)J data file in provided directory
}
