% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-vdj.R
\name{.identify_paired}
\alias{.identify_paired}
\title{Identify clonotypes with paired chains}
\usage{
.identify_paired(df_in, vdj_class)
}
\arguments{
\item{df_in}{data.frame containing V(D)J data formatted so each row
represents a single contig}

\item{vdj_class}{One of 'TCR' or 'BCR' to indicate type of data}
}
\value{
Input data.frame with paired column added
}
\description{
Identify clonotypes with paired chains
}
