% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vdj.R
\name{.load_vdj_data}
\alias{.load_vdj_data}
\title{Load V(D)J data}
\usage{
.load_vdj_data(vdj_dir, contig_file = "filtered_contig_annotations.csv")
}
\arguments{
\item{vdj_dir}{Directory containing the output from cellranger vdj. A vector
or named vector can be given to load data from several runs. If a named
vector is given, the cell barcodes will be prefixed with the provided names.
This mimics the behavior of the Read10X function found in the Seurat
package.}

\item{contig_file}{cellranger vdj output file containing data for each
contig annotation}
}
\value{
List containing one data.frame for each path provided to vdj_dir
}
\description{
Load V(D)J data
}
