% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-vdj.R
\name{.load_vdj_indels}
\alias{.load_vdj_indels}
\title{Load insertion/deletion information for each contig}
\usage{
.load_vdj_indels(vdj_dir, bam_file = "concat_ref.bam")
}
\arguments{
\item{vdj_dir}{Directory containing the output from cellranger vdj. A vector
or named vector can be given to load data from several runs. If a named
vector is given, the cell barcodes will be prefixed with the provided names.
This mimics the behavior of the Read10X function found in the Seurat
package.}

\item{bam_file}{bam file from cellranger vdj containing alignment data comparing
each contig with the germline reference}
}
\value{
List containing one data.frame for each path provided to vdj_dir
}
\description{
Load insertion/deletion information for each contig
}
