% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.merge_meta}
\alias{.merge_meta}
\title{Merge new meta.data with object}
\usage{
.merge_meta(input, meta, by = ".cell_id", pct_min = 25)
}
\arguments{
\item{input}{Object containing single cell data}

\item{meta}{meta.data to merge with object}

\item{by}{Columns to use for merging}

\item{pct_min}{If the percent overlap is less than pct_min, and error is
thrown}
}
\value{
Object with added meta.data
}
\description{
Merge new meta.data with object
}
