% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.set_lims}
\alias{.set_lims}
\title{Set min and max values for column}
\usage{
.set_lims(df_in, ft, mn = NULL, mx = NULL)
}
\arguments{
\item{df_in}{Input data.frame}

\item{ft}{Name of column containing feature values}

\item{mn}{Minimum percent rank}

\item{mx}{Maximum percent rank}
}
\value{
data.frame with modified feature values
}
\description{
Set min and max values for column
}
