% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.set_lvls}
\alias{.set_lvls}
\title{Set column levels}
\usage{
.set_lvls(df_in, clmn, lvls)
}
\arguments{
\item{df_in}{data.frame}

\item{clmn}{Column to modify}

\item{lvls}{Levels}
}
\value{
data.frame
}
\description{
Set column levels
}
