% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.split_vdj}
\alias{.split_vdj}
\title{Split V(D)J meta.data columns into vectors}
\usage{
.split_vdj(
  df_in,
  sep = ";",
  sep_cols,
  expand = FALSE,
  coerce_cols = c(umis = "numeric", reads = "numeric", cdr3_length = "numeric",
    cdr3_nt_length = "numeric", productive = "logical", full_length = "logical")
)
}
\arguments{
\item{df_in}{data.frame to modify}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{sep_cols}{Columns to split based on sep}

\item{expand}{Should columns be unnested after splitting into vectors}

\item{coerce_cols}{Named vector specifying columns that should be coerced to
a new type. The vector names should be the name of each column,
e.g., c(umis = "numeric")}
}
\value{
data.frame
}
\description{
Split V(D)J meta.data columns into vectors
}
