% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-vdj.R
\name{.subset_cells}
\alias{.subset_cells}
\alias{.subset_cells.default}
\alias{.subset_cells.data.frame}
\title{Subset object so it only contains given cells}
\usage{
.subset_cells(input, cells)

\method{.subset_cells}{default}(input, cells)

\method{.subset_cells}{data.frame}(input, cells)
}
\arguments{
\item{input}{Single cell object or data.frame. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{cells}{Vector with cell barcodes to use for subsetting}
}
\value{
Subsetted object
}
\description{
Subset object so it only contains given cells
}
