% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.to_tibble}
\alias{.to_tibble}
\title{Convert to tibble}
\usage{
.to_tibble(input, row_col)
}
\arguments{
\item{input}{Object to coerce to tibble}

\item{row_col}{Name of new column to store rownames}
}
\value{
tibble with rownames added to new column
}
\description{
The point of this function is to handle situations where the name passed to
rownames already exists in the data.frame. This occurs when running
.get_meta() multiple times on the same object. The default behavior of
as_tibble is to throw an error. This function just returns the tibble
without adding a new rowname column.
}
