% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_clonotypes}
\alias{identify_clonotypes}
\title{Identify clonotypes}
\usage{
identify_clonotypes(
  sobj_in,
  vdj_dir,
  csv_file = "enclone_output.csv",
  prefix = "",
  cell_prefix = "",
  overwrite = FALSE
)
}
\arguments{
\item{sobj_in}{Seurat object}

\item{vdj_dir}{cellranger vdj output directories. If a vector of multiple
paths is provided, an equal number of cell prefixes must also be provided.
If a named vector is given, the names will be used to prefix each cell
barcode.}

\item{csv_file}{Output file to write enclone results}

\item{prefix}{Prefix to add to new meta.data columns}

\item{cell_prefix}{Prefix to add to cell barcodes}

\item{overwrite}{If csv_file already exists, should the file be overwritten?
If csv_file already exists and overwrite = FALSE, the existing file will be
loaded. To re-run enclone, set overwrite = TRUE.}
}
\value{
Seurat object
}
\description{
Identify clonotypes using enclone from 10x Genomics
(https://10xgenomics.github.io/enclone)
}
