% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-gene-usage.R
\name{old_calc_gene_usage}
\alias{old_calc_gene_usage}
\title{old version}
\usage{
old_calc_gene_usage(
  input,
  data_cols,
  cluster_col = NULL,
  chain = NULL,
  chain_col = global$chain_col,
  sep = global$sep
)
}
\description{
old version
}
