% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot clonotype abundance}
\usage{
plot_abundance(
  sobj_in,
  clonotype_col = NULL,
  cluster_col = NULL,
  yaxis = "percent",
  plot_colors = NULL,
  plot_lvls = NULL,
  label_col = NULL,
  n_labels = 2,
  label_aes = list(),
  abundance_col = NULL,
  ...
)
}
\arguments{
\item{sobj_in}{Seurat object containing V(D)J data}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating clonotype abundance}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{yaxis}{Units to plot on the y-axis, either "frequency" or "percent"}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{label_col}{meta.data column containing labels to use for plot}

\item{n_labels}{Number of clonotypes to label}

\item{label_aes}{Named list providing additional label aesthetics (color, size, etc.)}

\item{abundance_col}{meta.data column containing pre-calculated abundances}

\item{...}{Additional arguments to pass to geom_line}
}
\value{
ggplot object
}
\description{
Plot clonotype abundance
}
