% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cdr3_length}
\alias{plot_cdr3_length}
\title{Plot CDR3 lengths}
\usage{
plot_cdr3_length(
  input,
  length_col = "cdr3_length",
  cluster_col = NULL,
  chain = NULL,
  type = "histogram",
  yaxis = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  chain_col = "chains",
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{length_col}{meta.data column containing CDR3 lengths to plot}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when plotting CDR3 lengths}

\item{chain}{Chain to use for plotting CDR3 lengths, set to NULL to include
all chains}

\item{type}{Type of plot to create, can be 'histogram' or 'violin'}

\item{yaxis}{Units to use for y-axis when plotting histogram. Use
'frequency' to show the raw number of CDR3 sequences or 'percent' to show
the percentage of total CDR3 sequences.}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Plot CDR3 lengths
}
