% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cell_count}
\alias{plot_cell_count}
\title{Create bar graphs summarizing cell counts for clusters}
\usage{
plot_cell_count(
  sobj_in,
  x,
  fill_col = NULL,
  split_col = NULL,
  yaxis = "fraction",
  plot_colors = NULL,
  plot_levels = NULL,
  order_count = TRUE,
  n_label = TRUE,
  label_aes = list(),
  split_rows = 1,
  split_scales = "free_x",
  ...
)
}
\arguments{
\item{sobj_in}{Seurat object or data.frame containing data}

\item{x}{meta.data column containing clusters to use for creating bars}

\item{fill_col}{meta.data column to use for coloring bars}

\item{split_col}{meta.data column containing groups to use for spliting bars}

\item{yaxis}{Units to plot on the y-axis, either "fraction" or "count"}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_levels}{Character vector containing levels for ordering}

\item{order_count}{Order bar color by number of cells in each group}

\item{n_label}{Include label showing the number of cells represented by each
bar}

\item{label_aes}{Named list providing additional label aesthetics
(color, size, etc.)}

\item{split_rows}{The number of facet rows. Use this argument if split_col
is specified}

\item{split_scales}{If split_col is used, this argument passes a scales
specification to facet_wrap. Can be "fixed", "free", "free_x", or "free_y"}

\item{...}{Additional arguments to pass to geom_bar}
}
\value{
ggplot object
}
\description{
Create bar graphs summarizing cell counts for clusters
}
