% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plot repertoire diversity}
\usage{
plot_diversity(
  sobj_in,
  clonotype_col = "clonotype_id",
  cluster_col = NULL,
  method = "simpson",
  plot_colors = NULL,
  plot_levels = NULL,
  ...
)
}
\arguments{
\item{sobj_in}{Seurat object containing V(D)J data or data.frame containing
diversity values to plot}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating clonotype abundance}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{method}{Method to use for calculating diversity}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_levels}{Character vector containing levels for ordering}

\item{...}{Additional arguments to pass to geom_col}
}
\value{
ggplot object
}
\description{
Plot repertoire diversity
}
