% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frequency.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot frequency of cell groups present in object}
\usage{
plot_frequency(
  input,
  data_col,
  cluster_col = NULL,
  group_col = NULL,
  method = "bar",
  stack = NULL,
  units = "percent",
  top = NULL,
  other_label = "other",
  plot_colors = NULL,
  plot_lvls = NULL,
  na_color = "grey80",
  trans = "identity",
  show_points = TRUE,
  show_zeros = TRUE,
  n_label = NULL,
  p_label = c(value = 0.05),
  p_method = NULL,
  p_file = NULL,
  label_params = list(),
  ...,
  per_chain = FALSE,
  chain = NULL,
  chain_col = global$chain_col,
  sep = global$sep
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing cell labels to use for
calculating frequency, e.g. cell types, isotypes etc.}

\item{cluster_col}{meta.data column containing cluster IDs (or patients,
treatment conditions, etc.) to use when calculating frequency. Calculations
will be performed separately for each cluster.}

\item{group_col}{meta.data column to use for grouping cluster IDs present in
cluster_col. This is useful when there are multiple replicates or patients
for each treatment condition.}

\item{method}{Method to use for plotting when \code{group_col} is provided,
possible values are 'bar' or 'boxplot'}

\item{stack}{If \code{TRUE}, stacked bargraphs will be generated, otherwise grouped
bargraphs will be generated}

\item{units}{Units to plot on the y-axis, either 'frequency' or 'percent'}

\item{top}{To only show the top cell groups present in \code{data_col}, provide
one of the following, all other cells will be labeled using the value
provided to the \code{other_label} argument. If \code{NULL} this will be automatically
set.
\itemize{
\item Integer specifying the number of top groups to show
\item Vector specifying the names of cell groups to show
}}

\item{other_label}{Label to use for 'other' cells when \code{top} is specified, if
\code{NULL} all cell groups present in data_col will be displayed on the plot.}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Levels to use for ordering clusters or groups}

\item{na_color}{Color to use for missing values}

\item{trans}{Transformation to use for plotting data, e.g. 'log10'. By
default values are not transformed, refer to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}
for more options. Values can only be transformed when stack is \code{FALSE}}

\item{show_points}{If \code{TRUE} data points will be shown on boxplots, the point
size can be adjusted using the \code{point.size} parameter}

\item{show_zeros}{If \code{TRUE} cell labels that are missing from a cluster will
still be shown on the plot}

\item{n_label}{Location on plot where n label should be added, this can be
any combination of the following:
\itemize{
\item 'corner', display the total number of cells plotted in the top right
corner, the position of the label can be modified by passing \code{x} and \code{y}
specifications with the \code{label_params} argument
\item 'axis', display the number of cells plotted for each group shown on the
x-axis
\item 'legend', display the number of cells plotted for each group shown in the
plot legend
\item 'none', do not display the number of cells plotted
}}

\item{p_label}{Specification indicating how p-values should be labeled on
plot, this can one of the following:
\itemize{
\item 'none', do not display p-values
\item 'all', show p-values for all groups
\item A named vector providing p-value cutoffs and labels to display,
e.g. \code{c('*' = 0.05, '**' = 0.01, '***' = 0.001)}. The keyword 'value' can
be used to display the p-value for those less than a certain cutoff,
e.g. \code{c(value = 0.05, ns = 1.1)} will show significant p-values, all others
will be labeled 'ns'.
}}

\item{p_method}{Method to use for calculating p-values, by default when
comparing two groups a t-test will be used.
When comparing more than two groups the Kruskal-Wallis test will be used.
With the exception of the edgeR method, p-values are adjusted for
multiple testing using Bonferroni correction. Possible methods include:
\itemize{
\item 't', two sample t-test performed with \code{stats::t.test()}
\item 'wilcox', Wilcoxon rank sum test performed with \code{stats::wilcox.test()}
\item 'kruskal', Kruskal-Wallis test performed with \code{stats::kruskal.test()}
\item 'edgeR', differential abundance calculated with the edgeR package
}}

\item{p_file}{File path to save table containing p-values for each
comparison.}

\item{label_params}{Named list providing additional parameters to modify
n label aesthetics, e.g. list(size = 4, color = "red")}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}

\item{per_chain}{If \code{TRUE} the frequency of each per-chain value will be
calculated. If \code{FALSE} per-chain data will not be parsed and the values
present in \code{data_col} will be used as is.}

\item{chain}{Chain(s) to use for calculating frequency. Set to \code{NULL} to
include all chains.}

\item{chain_col}{meta.data column(s) containing chains for each cell}

\item{sep}{Separator for storing per-chain data}
}
\value{
ggplot object
}
\description{
Plot the frequency of each cell label present in the provided meta.data
column. This is useful for comparing the proportion of cells belonging to
different samples, cell types, clonotypes, isotypes, etc.
}
\examples{
# Plot frequency of different isotypes
plot_frequency(
  vdj_so,
  data_col = "isotype"
)

# Plot frequency separately for cell clusters
plot_frequency(
  vdj_sce,
  data_col    = "isotype",
  cluster_col = "orig.ident"
)

# Create grouped bar graphs
plot_frequency(
  vdj_sce,
  data_col    = "isotype",
  cluster_col = "orig.ident",
  stack       = FALSE
)

# Plot number of cells on the y-axis
plot_frequency(
  vdj_so,
  data_col    = "seurat_clusters",
  cluster_col = "orig.ident",
  units       = "frequency"
)

}
\seealso{
\code{\link[=calc_frequency]{calc_frequency()}}, \code{\link[=plot_clone_frequency]{plot_clone_frequency()}}
}
