% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-gene-usage.R
\name{plot_gene_usage}
\alias{plot_gene_usage}
\title{Plot V(D)J gene usage}
\usage{
plot_gene_usage(
  input,
  gene_cols,
  cluster_col = NULL,
  group_col = NULL,
  chain = NULL,
  type = "bar",
  plot_colors = NULL,
  vdj_genes = NULL,
  n_genes = 20,
  plot_lvls = names(plot_colors),
  yaxis = "percent",
  chain_col = "chains",
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{gene_cols}{meta.data column containing genes for each clonotype,
provide a vector with two column names to plot paired usage of genes}

\item{cluster_col}{meta.data column containing cell clusters to use for
calculating gene usage}

\item{group_col}{meta.data column to use for grouping cluster IDs present in
cluster_col. This is useful when there are multiple replicates or patients
for each treatment condition.}

\item{chain}{Chain to use for calculating gene usage, set to NULL to include
all chains}

\item{type}{Type of plot to create, possible values are:
\itemize{
\item 'bar', bargraphs can only be plotted when a single column is passed to the
gene_cols argument
\item 'heatmap', heatmap implemented with ComplexHeatmap::Heatmap()
\item 'circos', circos plot implemented with circlize::chordDiagram().
to create a circos plot, two columns must be passed to the gene_cols
argument
}}

\item{plot_colors}{Character vector containing colors to use for plot. If a
bar graph is created this will specify how to color cell clusters. For a
heatmap, these colors will be used to generate the color gradient.}

\item{vdj_genes}{V(D)J genes to plot, if NULL the top genes will be shown}

\item{n_genes}{Number of top genes to plot based on usage. If cluster_col is
provided, top genes will be identified for each cluster.}

\item{plot_lvls}{Levels to use for ordering clusters}

\item{yaxis}{Units to plot on the y-axis, either 'frequency' or 'percent'}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Plot V(D)J gene usage
}
\examples{
# Plot V(D)J segment usage for all cells
plot_gene_usage(
  vdj_so,
  gene_cols = "v_gene"
)

# Plot gene usage separately for cell clusters
plot_gene_usage(
  vdj_sce,
  gene_cols = "v_gene",
  cluster_col = "orig.ident"
)

# Plot gene usage for a specific chain
plot_gene_usage(
  vdj_so,
  gene_cols = "v_gene",
  chain = c("IGH", "IGK")
)

# Plot gene usage for a specific chain
plot_gene_usage(
  vdj_sce,
  gene_cols = "v_gene",
  chain = c("IGH", "IGK")
)

# Create a heatmap
plot_gene_usage(
  vdj_so,
  gene_cols = "v_gene",
  type = "heatmap"
)

# Plot paired usage of V(D)J segments
plot_gene_usage(
  vdj_sce,
  gene_cols = c("v_gene", "j_gene"),
)

# Specify colors to use for each cell cluster
plot_gene_usage(
  vdj_so,
  gene_cols = "v_gene",
  cluster_col = "orig.ident",
  plot_colors = c(avid_2 = "blue", avid_1 = "green")
)

# Specify order to use for plotting cell clusters
plot_gene_usage(
  vdj_sce,
  gene_cols = "v_gene",
  cluster_col = "orig.ident",
  plot_lvls = c("avid_2", "avid_1")
)

# Specify certain V(D)J genes to include in plot
plot_gene_usage(
  vdj_so,
  gene_cols = "v_gene",
  vdj_genes = c("IGKV5-43", "IGLV1", "IGHV1-64")
)

# Specify the number of top V(D)J genes to include in plot
plot_gene_usage(
  vdj_sce,
  gene_cols = "v_gene",
  n_genes = 10
)

# Plot the frequency of each V(D)J segment instead of percent
plot_gene_usage(
  vdj_so,
  gene_cols = "v_gene",
  yaxis = "frequency"
)

}
