% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{plot_mds}
\alias{plot_mds}
\title{Create MDS plot}
\usage{
plot_mds(
  input,
  data_col,
  cluster_col,
  method = abdiv::jaccard,
  chain = NULL,
  chain_col = "chains",
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  label_points = TRUE,
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{method}{Method to use for comparing clusters, possible values are:
\itemize{
\item A function that takes two numeric vectors containing counts for each
clonotype in the object, such as most beta diversity functions provided by
the abdiv package. This will generate a heatmap.
}}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Levels to use for ordering clusters}

\item{label_points}{Label points on plot}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to (\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}}
}
\description{
Perform multidimensional scaling and plot results
}
\seealso{
\code{\link[=calc_mds]{calc_mds()}}
}
