% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vdj.R
\name{plot_numerical}
\alias{plot_numerical}
\alias{plot_histogram}
\alias{plot_violin}
\title{Plot numerical single-cell data}
\usage{
plot_histogram(
  input,
  data_col,
  cluster_col = NULL,
  group_col = NULL,
  method = "histogram",
  top = NULL,
  other_label = "other",
  units = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  trans = "identity",
  panel_nrow = NULL,
  panel_scales = "fixed",
  na_color = "grey80",
  n_label = NULL,
  label_params = list(),
  ...,
  per_chain = FALSE,
  chain = NULL,
  chain_col = global$chain_col,
  summary_fn = mean,
  sep = global$sep
)

plot_violin(
  input,
  data_col,
  cluster_col = NULL,
  group_col = NULL,
  method = "violin",
  top = NULL,
  other_label = "other",
  plot_colors = NULL,
  plot_lvls = NULL,
  trans = "identity",
  panel_nrow = NULL,
  panel_scales = "free_x",
  na_color = "grey80",
  n_label = NULL,
  label_params = list(),
  ...,
  per_chain = FALSE,
  chain = NULL,
  chain_col = global$chain_col,
  summary_fn = mean,
  sep = global$sep
)
}
\arguments{
\item{input}{Single cell object or data.frame, if a
data.frame is provided, cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing data to plot}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells for plotting}

\item{group_col}{meta.data column to use for grouping clusters into separate
panels}

\item{method}{Method to use for plotting, possible values include,
'histogram', 'density', 'boxplot', or 'violin'}

\item{top}{To only show the top cell groups, provide
one of the following, all other cells will be labeled using the value
provided to the \code{other_label} argument. If \code{NULL} this will be automatically
set.
\itemize{
\item Integer specifying the number of top groups to show
\item Vector specifying the names of cell groups to show
}}

\item{other_label}{Label to use for 'other' cells when \code{top} is specified, if
\code{NULL} all cell groups will be shown.}

\item{units}{Units to use for y-axis when method is 'histogram'. Use
'frequency' to show the number of values or 'percent' to show the percentage
of total values.}

\item{plot_colors}{Character vector specifying colors to use for cell
clusters specified by cluster_col. When cluster_col is \code{NULL}, plot colors
can be directly modified with the ggplot2 parameters \code{color} and \code{fill},
e.g. \code{fill = "red"}, \code{color = "black"}}

\item{plot_lvls}{Character vector containing order to use for plotting cell
clusters specified by cluster_col}

\item{trans}{Transformation to use when plotting data, e.g. 'log10'. By
default values are not transformed, refer to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}
for more options.}

\item{panel_nrow}{The number of rows to use for arranging plot panels}

\item{panel_scales}{Should scales for plot panels be fixed or free. This
passes a scales specification to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, can be 'fixed',
'free', 'free_x', or 'free_y'. 'fixed' will cause panels to share the same
scales.}

\item{na_color}{Color to use for missing values. If plotting V(D)J data,
cells lacking data will be plotted as \code{NA}s.}

\item{n_label}{Location on plot where n label should be added, this can be
any combination of the following:
\itemize{
\item 'corner', display the total number of cells plotted in the top right
corner, the position of the label can be modified by passing \code{x} and \code{y}
specifications with the \code{label_params} argument
\item 'axis', display the number of cells plotted for each group shown on the
x-axis
\item 'legend', display the number of cells plotted for each group shown in the
plot legend
\item 'none', do not display the number of cells plotted
}}

\item{label_params}{Named list providing additional parameters to modify
n label aesthetics, e.g. list(size = 4, color = "red")}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}

\item{per_chain}{If \code{TRUE} values will be plotted for each chain, i.e. each
data point represents a chain. If \code{FALSE} values will be summarized for each
cell using \code{summary_fn} before plotting, i.e. each data point represents a
cell.}

\item{chain}{Chain(s) to use for filtering data before plotting. If NULL
data will not be filtered based on chain.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{summary_fn}{Function to use for summarizing values when \code{per_chain} is
\code{FALSE}, can be either a function, e.g. \code{mean}, or a purrr-style
lambda, e.g. \code{~ mean(.x, na.rm = TRUE)} where \code{.x} refers to the column. If
\code{NULL}, the mean will be calculated.}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}
}
\value{
ggplot object
}
\description{
Visualize numerical single-cell data by creating a histogram, density plot,
violin plots, or boxplots. When plotting V(D)J data, values can be plotted
separately for each chain or summarized and plotted for each cell.
}
\examples{
# Create histogram
plot_histogram(
  vdj_so,
  data_col    = "reads",
  cluster_col = "orig.ident",
  trans       = "log10"
)

# Create violin plots
plot_violin(
  vdj_sce,
  data_col    = "nFeature_RNA",
  cluster_col = "orig.ident"
)

}
\seealso{
\code{\link[=summarize_vdj]{summarize_vdj()}} for more examples on how per-chain data can be
summarized for each cell
}
