% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot repertoire overlap}
\usage{
plot_overlap(
  obj_in,
  clonotype_col = NULL,
  cluster_col = NULL,
  plot_colors = NULL,
  ...
)
}
\arguments{
\item{obj_in}{Seurat object containing VDJ data or matrix}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{plot_colors}{Character vector containing colors to use for plotting}

\item{...}{Additional arguments to pass to geom_tile}
}
\value{
ggplot object
}
\description{
Plot repertoire overlap
}
