% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-diversity.R
\name{plot_rarefaction}
\alias{plot_rarefaction}
\title{Plot rarefaction curves}
\usage{
plot_rarefaction(
  input,
  data_col,
  cluster_col = NULL,
  method = "richness",
  n_boots = 50,
  chain = NULL,
  chain_col = "chains",
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  panel_nrow = NULL,
  panel_scales = "free",
  ci_alpha = 0.15,
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing values to use for
calculating diversity}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{method}{Method to use for calculating diversity, available methods
include:
\itemize{
\item 'richness', species richness, this is equivalent to the calculation
performed by \code{\link[abdiv:richness]{abdiv::richness()}}
\item 'shannon', the exponential of Shannon entropy
\item 'invsimpson', the inverse Simpson index, this is equivalent to the
calculation performed by \code{\link[abdiv:simpson]{abdiv::invsimpson()}}
}}

\item{n_boots}{Number of bootstrap replicates for calculating standard error,
if n_boots is 0 this will be skipped.}

\item{chain}{Chain to use for calculating diversity. To calculate diversity
for a single chain, the column passed to the data_col argument must contain
per-chain data such as CDR3 sequences. Set to NULL to include all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{panel_nrow}{The number of rows to use for arranging plot panels}

\item{panel_scales}{Should scales for plot panels be fixed or free. This
passes a scales specification to ggplot2::facet_wrap, can be 'fixed', 'free',
'free_x', or 'free_y'. 'fixed' will cause panels to share the same scales.}

\item{ci_alpha}{Transparency to use when plotting 95\% confidence interval}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill,
linetype, etc.}
}
\value{
ggplot object
}
\description{
This method involves calculating species diversity for different sized
samples generated by randomly downsampling each cluster. By default the
bootstrapped 95\% confidence interval will also be plotted.
}
\examples{
# Plot rarefaction curve using all cells
plot_rarefaction(
  vdj_so,
  data_col = "clonotype_id"
)

# Specify method to use for calculating repertoire diversity
plot_rarefaction(
  vdj_sce,
  data_col = "clonotype_id",
  method   = "shannon"
)

# Plot separate curves for each cell cluster
plot_rarefaction(
  vdj_so,
  data_col    = "clonotype_id",
  cluster_col = "orig.ident"
)

# Specify colors to use for each cell cluster
plot_rarefaction(
  vdj_so,
  data_col    = "clonotype_id",
  cluster_col = "orig.ident",
  plot_colors = c(avid_1 = "green", avid_2 = "purple")
)

# Specify order to use for plotting cell clusters
plot_rarefaction(
  vdj_sce,
  data_col    = "clonotype_id",
  cluster_col = "orig.ident",
  plot_lvls   = c("avid_2", "avid_1")
)

}
\seealso{
\code{\link[=calc_diversity]{calc_diversity()}}, \code{\link[=plot_diversity]{plot_diversity()}}
}
