% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-plot-vdj.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Create scatter plot}
\usage{
plot_scatter(
  input,
  data_col = NULL,
  x = "UMAP_1",
  y = "UMAP_2",
  group_col = NULL,
  summary_fn = NULL,
  chain = NULL,
  chain_col = global$chain_col,
  top = NULL,
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  trans = "identity",
  min_q = NULL,
  max_q = NULL,
  panel_nrow = NULL,
  panel_scales = "fixed",
  na_color = "grey80",
  other_label = "other",
  n_label = NULL,
  label_params = list(),
  sep = global$sep,
  ...
)
}
\description{
Create scatter plot
}
