% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-seqs.R
\name{plot_seq_motifs}
\alias{plot_seq_motifs}
\title{Create sequence logos for clusters}
\usage{
plot_seq_motifs(
  input,
  data_col = "cdr3",
  cluster_col = NULL,
  chain,
  plot_colors = NULL,
  plot_lvls = NULL,
  chain_col = "chains",
  width = 0.75,
  align_end = "5",
  facet_rows = NULL,
  sep = ";",
  ...
)
}
\description{
Create sequence logos for clusters
}
