% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_usage}
\alias{plot_usage}
\title{Plot VDJ gene usage}
\usage{
plot_usage(
  sobj_in,
  gene_col,
  cluster_col = NULL,
  plot_colors = NULL,
  plot_genes = NULL,
  n_genes = NULL,
  clust_levels = NULL,
  line_color = NULL,
  ...
)
}
\arguments{
\item{sobj_in}{Seurat object containing VDJ data}

\item{gene_col}{meta.data column containing genes used for each clonotype}

\item{cluster_col}{meta.data column containing cell clusters to use when
calculating gene usage}

\item{plot_colors}{Character vector containing colors to use for plotting}

\item{plot_genes}{Character vector of genes to plot}

\item{n_genes}{Number of top genes to plot based on average usage}

\item{clust_levels}{Levels to use for ordering clusters}

\item{line_color}{Color of line used to separate clusters}

\item{...}{Additional arguments to pass to geom_tile}
}
\value{
ggplot object
}
\description{
Plot VDJ gene usage
}
