% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_vdj}
\alias{plot_vdj}
\alias{plot_vdj_reads}
\alias{plot_cdr3_length}
\alias{plot_vdj_indels}
\title{Plot continuous V(D)J data}
\usage{
plot_vdj(
  input,
  data_cols,
  per_cell = FALSE,
  summary_fn = mean,
  cluster_col = NULL,
  chain = NULL,
  type = "histogram",
  yaxis = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  alpha = 0.5,
  log_trans = FALSE,
  chain_col = "chains",
  sep = ";",
  ...
)

plot_vdj_reads(
  input,
  data_cols = c("reads", "umis"),
  cluster_col = NULL,
  chain = NULL,
  type = "violin",
  yaxis = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  chain_col = "chains",
  sep = ";",
  ...
)

plot_cdr3_length(
  input,
  data_cols = "cdr3_length",
  cluster_col = NULL,
  chain = NULL,
  type = "histogram",
  yaxis = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  chain_col = "chains",
  sep = ";",
  ...
)

plot_vdj_indels(
  input,
  data_cols = c("n_insertion", "n_deletion", "n_mismatch"),
  cluster_col = NULL,
  chain = NULL,
  type = "violin",
  yaxis = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  chain_col = "chains",
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, cell barcodes should be stored as row names.}

\item{data_cols}{meta.data column(s) containing continuous V(D)J data to
plot}

\item{per_cell}{Should values be plotted per cell, i.e. each data point
would represent one cell. If TRUE, values will be summarized for each cell
using summary_fn. If FALSE, values will be plotted per chain.}

\item{summary_fn}{Function to use for summarizing values for each cell, e.g.
specifying stats::median will plot the median value per cell}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when plotting CDR3 lengths}

\item{chain}{Chain(s) to use for plotting, set to NULL to include all chains}

\item{type}{Type of plot to create, can be 'histogram', 'density',
'boxplot', or 'violin'}

\item{yaxis}{Units to use for y-axis when plotting histogram. Use
'frequency' to show the raw number of values or 'percent' to show
the percentage of total values.}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{alpha}{Color transparency}

\item{log_trans}{If TRUE, axis will be log10 transformed}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{...}{Additional arguments to pass to ggplot geom}
}
\value{
ggplot object
}
\description{
Plot continuous V(D)J data
}
