% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-usage.R
\name{plot_vdj_usage}
\alias{plot_vdj_usage}
\title{Plot V(D)J gene usage}
\usage{
plot_vdj_usage(
  input,
  gene_cols,
  cluster_col = NULL,
  chain = NULL,
  type = "bar",
  plot_colors = NULL,
  vdj_genes = NULL,
  n_genes = 50,
  plot_lvls = NULL,
  yaxis = "percent",
  chain_col = "chains",
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{gene_cols}{meta.data column containing genes for each clonotype,
provide a vector with two column names to plot paired usage of genes}

\item{cluster_col}{meta.data column containing cell clusters to use for
calculating gene usage}

\item{chain}{Chain to use for calculating gene usage, set to NULL to include
all chains}

\item{type}{Type of plot to create, can be either 'heatmap' or 'bar'. If
multiple columns are provided to gene_cols, a heatmap is generated.}

\item{plot_colors}{Character vector containing colors to use for plot. If a
bar graph is created this will specify how to color cell clusters. For a
heatmap, these colors will be used to generate the color gradient.}

\item{vdj_genes}{Character vector containing genes to plot}

\item{n_genes}{Number of top genes to plot based on usage. If cluster_col is
provided, top genes will be identified for each cluster.}

\item{plot_lvls}{Levels to use for ordering clusters}

\item{yaxis}{Units to plot on the y-axis, either 'frequency' or 'percent'}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Plot V(D)J gene usage
}
\examples{
# Plot V(D)J segment usage for all cells
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene"
)

# Plot gene usage separately for cell clusters
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  cluster_col = "orig.ident"
)

# Plot gene usage for a specific chain
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  chain = c("IGH", "IGK")
)

# Plot gene usage for a specific chain
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  chain = c("IGH", "IGK")
)

# Create a heatmap
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  type = "heatmap"
)

# Plot paired usage of V(D)J segments
plot_vdj_usage(
  vdj_so,
  gene_cols = c("v_gene", "j_gene"),
)

# Specify colors to use for each cell cluster
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  cluster_col = "orig.ident",
  plot_colors = c(avid_2 = "blue", avid_1 = "green")
)

# Specify order to use for plotting cell clusters
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  cluster_col = "orig.ident",
  plot_lvls = c("avid_2", "avid_1")
)

# Specify certain V(D)J genes to include in plot
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  vdj_genes = c("IGKV5-43", "IGLV1", "IGHV1-64", "IGHV1-11")
)

# Specify the number of top V(D)J genes to include in plot
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  n_genes = 10
)

# Plot the frequency of each V(D)J segment instead of percent
plot_vdj_usage(
  vdj_so,
  gene_cols = "v_gene",
  yaxis = "frequency"
)

# Using plot_vdj_usage outside of Seurat
# SingleCellExperiment objects or data.frames containing V(D)J data are also
# compatible. If a data.frame is provided, cell barcodes should be stored as
# row names.
plot_vdj_usage(vdj_sce, gene_cols = "v_gene")

df <- vdj_so@meta.data

plot_vdj_usage(df, gene_cols = "v_gene")

}
