% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{set_val_limits}
\alias{set_val_limits}
\title{Set min and max values for column}
\usage{
set_val_limits(df_in, feat_col, lim, op)
}
\arguments{
\item{df_in}{Input data.frame}

\item{feat_col}{Name of column containing feature values}

\item{lim}{The value cutoff}

\item{op}{The operator to use for comparing values with lim
(either "<" or ">")}
}
\value{
data.frame with modified feature values
}
\description{
Set min and max values for column
}
