% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{so}
\alias{so}
\title{AVID-seq data for mouse splenocytes}
\format{
A Seurat object with 12334 features, 7345 cells, and 3 assays
}
\usage{
so
}
\description{
A Seurat object containing AVID-seq data for mouse splenocytes
}
\keyword{datasets}
