% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{splen_meta}
\alias{splen_meta}
\title{Meta data for downsampled splenocyte data}
\format{
A data.frame containing meta data for 500 cells
}
\usage{
splen_meta
}
\value{
Meta data for downsampled splenocyte data
}
\description{
Meta data for downsampled splenocyte data
}
\keyword{datasets}
