% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_chains}
\alias{summarize_chains}
\title{Summarize values for chains}
\usage{
summarize_chains(
  input,
  data_cols = c("umis", "reads"),
  fn,
  chain_col = "chains",
  include_cols = NULL,
  sep = ";"
)
}
\arguments{
\item{input}{Object containing single cell data}

\item{data_cols}{meta.data columns to summarize}

\item{fn}{Function to use for summarizing data_cols}

\item{chain_col}{meta.data column(s) containing labels for each chain
expressed in the cell. These labels are used for grouping the summary
output. Set chain_col to NULL to group solely based on the cell barcodes.}

\item{include_cols}{Additional columns to include in the output data.frame}

\item{sep}{Separator used for storing per cell V(D)J data}
}
\value{
data.frame containing summary results
}
\description{
Summarize values present for each column provided to the data_cols argument.
For each cell, the function(s) provided will be applied to each unique label
in chain_col.
}
