% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_vdj}
\alias{summarize_vdj}
\title{Summarize V(D)J data for each cell}
\usage{
summarize_vdj(
  input,
  vdj_cols,
  fn = mean,
  chain = NULL,
  chain_col = "chains",
  sep = ";",
  col_names = "{.col}",
  return_df = FALSE
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{vdj_cols}{meta.data columns containing V(D)J data to summarize for
each cell}

\item{fn}{Function to apply to each of the selected columns, possible values
are: a function, e.g. mean; a purrr-style lambda, e.g. ~ mean(.x, na.rm =
TRUE). If NULL, unmodified meta.data is returned.}

\item{chain}{Chain to use for summarizing V(D)J data}

\item{chain_col}{meta.data column(s) containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{col_names}{A glue specification that describes how to name the output
columns, this can use {.col} to stand for the selected column name}

\item{return_df}{Return results as a data.frame. If FALSE, results will be
added to the input object.}
}
\value{
data.frame containing V(D)J data summarized for each cell
}
\description{
Summarize values present for each column provided to the vdj_cols argument.
The provided function will be applied to all values present for the cell.
}
