% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_all_args}
\alias{test_all_args}
\title{Helper to test all combinations of provided arguments}
\usage{
test_all_args(arg_lst, .fn, desc, chk, dryrun = FALSE)
}
\arguments{
\item{arg_lst}{Named list of arguments to test}

\item{.fn}{Function to test}

\item{desc}{Description to pass to test_that}

\item{chk}{Function or expression to using for testing}

\item{dryrun}{Do not run tests, just return table of arguments that will be
tested}
}
\value{
Output from test_that
}
\description{
Helper to test all combinations of provided arguments
}
