% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_so}
\alias{test_so}
\title{AVID-seq data for tests}
\format{
A Seurat object with 12334 features, 100 cells, and 3 assays
}
\usage{
test_so
}
\description{
A tiny Seurat object containing AVID-seq data used for tests
}
\keyword{datasets}
