% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_sce}
\alias{tiny_sce}
\title{A tiny SingleCellExperiment object}
\format{
A SingleCellExperiment object with 200 features, 200 cells, and 1
assay
}
\usage{
tiny_sce
}
\description{
A tiny SingleCellExperiment object
}
\keyword{datasets}
