% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdj_sce}
\alias{vdj_sce}
\title{A tiny SingleCellExperiment object with V(D)J data}
\format{
A SingleCellExperiment object with 200 features, 200 cells, and 1 assay
}
\usage{
vdj_sce
}
\description{
A tiny SingleCellExperiment object with V(D)J data
}
\keyword{datasets}
