% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdj_so}
\alias{vdj_so}
\title{A tiny Seurat object with V(D)J data}
\format{
A Seurat object with 200 features, 200 cells, and 1 assay
}
\usage{
vdj_so
}
\description{
A tiny Seurat object with V(D)J data
}
\keyword{datasets}
