% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abundance.R
\name{calc_abundance}
\alias{calc_abundance}
\alias{calc_abundance.default}
\alias{calc_abundance.Seurat}
\title{Calculate clonotype abundance}
\usage{
calc_abundance(input, ...)

\method{calc_abundance}{default}(
  input,
  clonotype_col = "cdr3_nt",
  cluster_col = NULL,
  prefix = "",
  ...
)

\method{calc_abundance}{Seurat}(
  input,
  clonotype_col = "cdr3_nt",
  cluster_col = NULL,
  prefix = "",
  return_seurat = TRUE
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{clonotype_col}{meta.data column containing clonotype IDs}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{prefix}{Prefix to add to new meta.data columns}

\item{return_seurat}{Return a Seurat object. If set to FALSE, a tibble
summarizing the results is returned.}
}
\value{
Seurat object with clonotype abundance added to meta.data
}
\description{
Calculate clonotype abundance
}
