% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_abundance}
\alias{calc_abundance}
\title{Calculate clonotype abundance}
\usage{
calc_abundance(
  sobj_in,
  clonotype_col,
  cluster_col = NULL,
  prefix = "",
  return_seurat = TRUE
)
}
\arguments{
\item{sobj_in}{Seurat object containing V(D)J data}

\item{clonotype_col}{meta.data column containing clonotype IDs}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{prefix}{Prefix to add to new meta.data columns}

\item{return_seurat}{Return a Seurat object. If set to FALSE, a tibble
summarizing the results is returned.}
}
\value{
Seurat object with clonotype abundance added to meta.data
}
\description{
Calculate clonotype abundance
}
