% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-abundance.R
\name{calc_abundance}
\alias{calc_abundance}
\title{Calculate clonotype abundance}
\usage{
calc_abundance(
  input,
  cluster_col = NULL,
  clonotype_col = "clonotype_id",
  prefix = "",
  return_df = FALSE
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{cluster_col}{meta.data column containing cluster IDs to use for}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating clonotype abundance}

\item{prefix}{Prefix to add to new columns}

\item{return_df}{Return results as a data.frame. If set to FALSE, results
will be added to the input object.}
}
\value{
Single cell object or data.frame with clonotype abundance metrics
}
\description{
Calculate clonotype abundance
}
