% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diversity.R
\name{calc_diversity}
\alias{calc_diversity}
\title{Calculate repertoire diversity}
\usage{
calc_diversity(
  input,
  cluster_col = NULL,
  method = abdiv::simpson,
  clonotype_col = "clonotype_id",
  prefix = "",
  return_df = FALSE
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating diversity. If cluster_col is omitted, diversity index will be
calculated for all clonotypes.}

\item{method}{Method to use for calculating diversity. A named list can also
be passed to use multiple methods. The names should specify names for the
output columns.}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating diversity}

\item{prefix}{Prefix to add to new columns}

\item{return_df}{Return results as a data.frame. If set to FALSE, results
will be added to the input object.}
}
\value{
Single cell object or data.frame with diversity metrics
}
\description{
Calculate repertoire diversity
}
