% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_diversity}
\alias{calc_diversity}
\title{Calculate receptor diversity (inverse Simpson Index)}
\usage{
calc_diversity(sobj_in, clonotype_col = "clonotype_id", cluster_col = NULL)
}
\arguments{
\item{sobj_in}{Seurat object}

\item{clonotype_col}{meta.data column containing clonotype ids}

\item{cluster_col}{meta.data column containing cluster ids used for calculating
diversity. If cluster_col is omitted, diversity index will be calculated for
all clonotypes}
}
\value{
Seurat object with inverse Simpson index added to meta.data
}
\description{
Calculate receptor diversity (inverse Simpson Index)
}
