% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_diversity}
\alias{calc_diversity}
\title{Calculate repertoire diversity}
\usage{
calc_diversity(
  sobj_in,
  clonotype_col,
  cluster_col = NULL,
  method = abdiv::simpson,
  prefix = "",
  return_seurat = TRUE
)
}
\arguments{
\item{sobj_in}{Seurat object}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating diversity}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating diversity. If cluster_col is omitted, diversity index will be
calculated for all clonotypes.}

\item{method}{Method to use for calculating diversity. Can pass also pass
a named list to use multiple methods.}

\item{prefix}{Prefix to add to new meta.data columns}

\item{return_seurat}{Return a Seurat object. If set to FALSE, the meta.data
table is returned.}
}
\value{
Seurat object with diversity index added to meta.data
}
\description{
Calculate repertoire diversity
}
