% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frequency.R
\name{calc_frequency}
\alias{calc_frequency}
\title{Calculate frequency of cell groups present in object}
\usage{
calc_frequency(
  input,
  data_col,
  cluster_col = NULL,
  prefix = paste0(data_col, "_"),
  return_df = FALSE
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing cell labels to use for
calculating frequency. To calculate clonotype frequencies, provide the column
containing clonotype IDs, to calculate isotype frequencies provide the column
containing cell isotypes. By default the clonotype_id is used for
calculations.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{prefix}{Prefix to add to new columns}

\item{return_df}{Return results as a data.frame. If set to FALSE, results
will be added to the input object.}
}
\value{
Single cell object or data.frame with clonotype frequencies
}
\description{
Calculate the frequency of each cell label present in the provided meta.data
column. This is useful for comparing the proportion of cells belonging to
different samples, cell types, clonotypes, isotypes, etc.
}
\examples{
# Calculate clonotype abundance using all cells
res <- calc_frequency(
  vdj_so,
  data_col = "clonotype_id"
)

head(slot(res, "meta.data"), 1)

# Group cells based on meta.data column before calculating abundance
res <- calc_frequency(
  vdj_sce,
  data_col = "clonotype_id",
  cluster_col = "orig.ident"
)

head(slot(res, "colData"), 1)

# Add a prefix to the new columns
# this is useful if multiple abundance calculations are stored in the
# meta.data
res <- calc_frequency(
  vdj_so,
  data_col = "clonotype_id",
  prefix = "bcr_"
)

head(slot(res, "meta.data"), 1)

# Return a data.frame instead of adding the results to the input object
res <- calc_frequency(
  vdj_sce,
  data_col = "clonotype_id",
  return_df = TRUE
)

head(res, 1)

}
\seealso{
\code{\link[=plot_frequency]{plot_frequency()}}, \code{\link[=plot_clone_frequency]{plot_clone_frequency()}}
}
