% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{calc_mds}
\alias{calc_mds}
\title{Perform multidimensional scaling}
\usage{
calc_mds(
  input,
  data_col,
  cluster_col,
  method = abdiv::jaccard,
  chain = NULL,
  chain_col = "chains",
  prefix = "",
  return_df = FALSE,
  sep = ";"
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing values to use for
calculating pairwise similarity between clusters, e.g. 'clonotype_id'}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating repertoire overlap}

\item{method}{Method to use for comparing clusters and calculating MDS
coordinates, available methods include:
\itemize{
\item 'jaccard', Jaccard dissimilarity index implemented with \code{\link[abdiv:jaccard]{abdiv::jaccard()}}
\item 'horn_morisita', Horn-Morisita index implemented with
\code{\link[abdiv:morisita]{abdiv::horn_morisita()}}
}}

\item{chain}{Chain to use for comparing clusters. To perform calculations
using a single chain, the column passed to the data_col argument must contain
per-chain data such as CDR3 sequences. Set to NULL to include all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{prefix}{Prefix to add to new columns}

\item{return_df}{Return results as a data.frame. If set to FALSE, results
will be added to the input object.}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}
}
\value{
Single cell object or data.frame with MDS coordinates
}
\description{
Calculate MDS coordinates based on a beta diversity metric.
}
\examples{
data(vdj_so, vdj_sce)

# Calculate MDS coordinates
res <- calc_mds(
  vdj_so,
  data_col = "clonotype_id",
  cluster_col = "isotype"
)

# Calculate MDS coordinates based on IGK CDR3 sequences
res <- calc_mds(
  vdj_sce,
  data_col    = "cdr3",
  cluster_col = "isotype",
  chain       = "IGK"
)

# Change the method used for calculating repertoire similarity
res <- calc_mds(
  vdj_sce,
  data_col    = "clonotype_id",
  cluster_col = "isotype",
  method      = "horn_morisita"
)

}
\seealso{
\code{\link[=plot_mds]{plot_mds()}}, \code{\link[=calc_similarity]{calc_similarity()}}, \code{\link[=plot_similarity]{plot_similarity()}},
\code{\link[MASS:isoMDS]{MASS::isoMDS()}}
}
