% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_similarity.R
\name{calc_similarity}
\alias{calc_similarity}
\alias{calc_similarity.default}
\alias{calc_similarity.Seurat}
\title{Calculate repertoire overlap}
\usage{
calc_similarity(input, ...)

\method{calc_similarity}{default}(
  input,
  clonotype_col = "cdr3_nt",
  cluster_col,
  method = abdiv::jaccard,
  prefix = NULL,
  return_mat = FALSE,
  ...
)

\method{calc_similarity}{Seurat}(
  input,
  clonotype_col = "cdr3_nt",
  cluster_col,
  method = abdiv::jaccard,
  prefix = NULL,
  return_mat = FALSE,
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{...}{Arguments passed to other methods}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{method}{Method to use for calculating similarity between clusters}

\item{prefix}{Prefix to add to new columns}

\item{return_mat}{Return a matrix with similarity values. If set to
FALSE, results will be added to the input object.}
}
\value{
Single cell object or data.frame with similarity values
}
\description{
Calculate repertoire overlap
}
