% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{calc_similarity}
\alias{calc_similarity}
\title{Calculate repertoire overlap}
\usage{
calc_similarity(
  input,
  cluster_col,
  method = abdiv::jaccard,
  clonotype_col = "clonotype_id",
  prefix = NULL,
  return_mat = FALSE
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating repertoire overlap}

\item{method}{Method to use for calculating similarity between clusters}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{prefix}{Prefix to add to new columns}

\item{return_mat}{Return a matrix with similarity values. If set to
FALSE, results will be added to the input object.}
}
\value{
Single cell object or data.frame with similarity values
}
\description{
Calculate repertoire overlap
}
\examples{
# Calculate repertoire overlap
res <- calc_similarity(
  vdj_so,
  clonotype_col = "clonotype_id",
  cluster_col = "orig.ident",
  method = abdiv::jaccard
)

head(res@meta.data, 1)

# Add a prefix to the new columns
# this is useful if multiple calculations are stored in the meta.data
res <- calc_similarity(
  vdj_sce,
  cluster_col = "orig.ident",
  prefix = "bcr_"
)

head(res@colData, 1)

# Return a matrix instead of adding the results to the input object
calc_similarity(
  vdj_so,
  cluster_col = "orig.ident",
  return_mat = TRUE
)

}
