% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{calc_similarity}
\alias{calc_similarity}
\title{Calculate repertoire overlap}
\usage{
calc_similarity(
  input,
  cluster_col,
  method = abdiv::jaccard,
  clonotype_col = "clonotype_id",
  prefix = NULL,
  return_mat = FALSE
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{method}{Method to use for calculating similarity between clusters}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{prefix}{Prefix to add to new columns}

\item{return_mat}{Return a matrix with similarity values. If set to
FALSE, results will be added to the input object.}
}
\value{
Single cell object or data.frame with similarity values
}
\description{
Calculate repertoire overlap
}
