% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_usage.R
\name{calc_usage}
\alias{calc_usage}
\alias{calc_usage.default}
\alias{calc_usage.Seurat}
\title{Calculate gene usage}
\usage{
calc_usage(input, ...)

\method{calc_usage}{default}(
  input,
  gene_cols,
  cluster_col = NULL,
  chain = NULL,
  chain_col = "chains",
  sep = ";"
)

\method{calc_usage}{Seurat}(
  input,
  gene_cols,
  cluster_col = NULL,
  chain = NULL,
  chain_col = "chains",
  sep = ";"
)
}
\arguments{
\item{input}{Seurat object containing V(D)J data}

\item{gene_cols}{meta.data column containing genes used for each clonotype}

\item{cluster_col}{meta.data column containing cell clusters to use when
calculating gene usage}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator to use for expanding gene_cols}
}
\value{
data.frame containing gene usage summary
}
\description{
Gene usage is calculated as the percent of cells that express each V(D)J
gene present in gene_cols. Cells that lack V(D)J data and have NAs present
in gene_cols are excluded from this calculation.
}
