% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_usage}
\alias{calc_usage}
\title{Calculate gene usage}
\usage{
calc_usage(
  sobj_in,
  gene_cols,
  cluster_col = NULL,
  chain = NULL,
  chain_col = "chains",
  sep = ";"
)
}
\arguments{
\item{sobj_in}{Seurat object containing V(D)J data}

\item{gene_cols}{meta.data column containing genes used for each clonotype}

\item{cluster_col}{meta.data column containing cell clusters to use when
calculating gene usage}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator to use for expanding gene_cols}
}
\value{
data.frame containing gene usage summary
}
\description{
Gene usage is calculated as the percent of cells that express each V(D)J
gene present in gene_cols. Cells that lack V(D)J data and have NAs present
in gene_cols are excluded from this calculation.
}
