% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_usage}
\alias{calc_usage}
\title{Calculate gene usage}
\usage{
calc_usage(sobj_in, gene_col, cluster_col = NULL)
}
\arguments{
\item{sobj_in}{Seurat object containing VDJ data}

\item{gene_col}{meta.data column containing genes used for each clonotype}

\item{cluster_col}{meta.data column containing cell clusters to use when
calculating gene usage}
}
\value{
data.frame containing gene usage summary
}
\description{
Calculate gene usage
}
