% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_usage}
\alias{calc_usage}
\title{Calculate gene usage}
\usage{
calc_usage(
  sobj_in,
  gene_col,
  cluster_col = NULL,
  chain = NULL,
  chain_col = NULL,
  sep = ";"
)
}
\arguments{
\item{sobj_in}{Seurat object containing V(D)J data}

\item{gene_col}{meta.data column containing genes used for each clonotype}

\item{cluster_col}{meta.data column containing cell clusters to use when
calculating gene usage}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator to use for expanding gene_col}
}
\value{
data.frame containing gene usage summary
}
\description{
Gene usage is calculated as the percent of cells that express each V(D)J
gene present in gene_col. Cells that lack V(D)J data and have an NA present
in gene_col are excluded from this calculation.
}
