% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-vdj.R
\name{define_clonotypes}
\alias{define_clonotypes}
\title{Define clonotypes based on V(D)J data}
\usage{
define_clonotypes(
  input,
  vdj_cols,
  clonotype_col = "clonotype_id",
  filter_chains = c("productive", "full_length"),
  sep = ";"
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{vdj_cols}{meta.data columns containing V(D)J data to use for defining
clonotypes}

\item{clonotype_col}{Name of column to use for storing clonotype IDs}

\item{filter_chains}{Columns to use for identifying productive full length
chains when defining clonotypes. If NULL, all chains are used for defining
clonotypes.}

\item{sep}{Separator used for storing per cell V(D)J data}
}
\value{
Single cell object or data.frame with added clonotype IDs
}
\description{
Define clonotypes based on V(D)J data
}
