% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djvdj-package.R
\docType{package}
\name{djvdj-package}
\alias{djvdj}
\alias{djvdj-package}
\title{djvdj: A Collection of Tools to Analyze AVID-seq Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
