% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djvdj-package.R
\docType{package}
\name{djvdj-package}
\alias{djvdj}
\alias{djvdj-package}
\title{djvdj: A collection of single-cell V(D)J tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

djvdj provides a single-cell V(D)J analysis workflow that integrates with popular single-cell data structures. This includes the ability to cluster CDR3 sequences and to quantify repertoire diversity, repertoire similarity, and gene segment usage. A variety of visualization methods are also provided including rarefaction curves, circos plots and UMAP projections.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rnabioco.github.io/djvdj/}
  \item \url{https://github.com/rnabioco/djvdj/}
  \item Report bugs at \url{https://github.com/rnabioco/djvdj/issues}
}

}
\author{
\strong{Maintainer}: Ryan Sheridan \email{ryan.sheridan@cuanschutz.edu} (\href{https://orcid.org/0000-0003-4012-3147}{ORCID})

Other contributors:
\itemize{
  \item Jay Hesselberth \email{jay.hesselberth@gmail.com} [contributor]
  \item Rui Fu \email{raysinensis@gmail.com} [contributor]
  \item Kent Riemondy \email{kent.riemondy@ucdenver.edu} [contributor]
  \item RNA Bioscience Initiative [funder]
}

}
\keyword{internal}
