% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-utils.R
\name{djvdj_theme}
\alias{djvdj_theme}
\alias{\%+replace\%}
\title{Theme for djvdj plotting functions}
\usage{
djvdj_theme(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  line_color = "grey85"
)
}
\arguments{
\item{base_size}{base font size in pts}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect element}

\item{line_color}{color for line elements}
}
\value{
ggplot theme
}
\description{
Theme for djvdj plotting functions
}
\examples{

plot_scatter(vdj_so, data_col = "seurat_clusters") +
  djvdj_theme()

}
