% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-vdj.R
\name{.classify_vdj}
\alias{.classify_vdj}
\title{Determine whether TCR or BCR data were provided}
\usage{
.classify_vdj(df_in, chain_col = "chains")
}
\arguments{
\item{df_in}{data.frame containing V(D)J data formatted so that each row
represents a single contig}

\item{chain_col}{Column in input data containing chain identity}
}
\value{
Character string indicating whether TCR or BCR data were provided
}
\description{
Determine whether TCR or BCR data were provided
}
