% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.convert_char}
\alias{.convert_char}
\title{Attempt to coerce character vector using provided function}
\usage{
.convert_char(x, Class)
}
\arguments{
\item{x}{Character vector to coerce}

\item{Class}{Name of the class to which x should be coerced}
}
\value{
Coerced vector
}
\description{
Attempt to coerce character vector using provided function
}
