% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.create_bars}
\alias{.create_bars}
\title{Create ggplot bar graph}
\usage{
.create_bars(
  df_in,
  x,
  y,
  .fill,
  plot_colors = NULL,
  y_ttl = y,
  ang = 45,
  hjst = 1,
  ...
)
}
\arguments{
\item{df_in}{data.frame}

\item{x}{Variable to plot on x-axis}

\item{y}{Variable to plot on y-axis}

\item{.fill}{Variable to use for fill color}

\item{plot_colors}{Vector of colors for plotting}

\item{y_ttl}{Title for y-axis}

\item{ang}{Angle of x-axis text}

\item{hjst}{Horizontal justification for x-axis text}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Create ggplot bar graph
}
