% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_vdj_cols}
\alias{.get_vdj_cols}
\title{Identify columns with V(D)J data}
\usage{
.get_vdj_cols(df_in, clone_col, cols_in, sep)
}
\arguments{
\item{df_in}{data.frame}

\item{clone_col}{Column containing clonotype IDs to use for identifying
columns with V(D)J data. If NULL all columns are used.}

\item{cols_in}{meta.data columns containing VDJ data to use for filtering.
If set to NULL (recommended) columns are automatically selected.}

\item{sep}{Separator to search for in columns}
}
\value{
list of vectors containing columns with V(D)J data and sep
}
\description{
Identify columns with V(D)J data
}
