% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vdj.R
\name{.merge_meta}
\alias{.merge_meta}
\title{Merge new meta.data with object}
\usage{
.merge_meta(input, meta, by = ".cell_id")
}
\arguments{
\item{input}{Object containing single cell data}

\item{meta}{meta.data to merge with object}

\item{by}{Columns to use for merging}
}
\value{
Object with added meta.data
}
\description{
Merge new meta.data with object
}
