% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.split_vdj}
\alias{.split_vdj}
\title{Split V(D)J meta.data columns into vectors}
\usage{
.split_vdj(
  df_in,
  sep = ";",
  sep_cols,
  num_cols = c("umis", "reads"),
  lgl_cols = c("productive", "full_length")
)
}
\arguments{
\item{df_in}{data.frame to modify}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{sep_cols}{Columns to split based on sep}

\item{num_cols}{Columns to convert to numeric}

\item{lgl_cols}{Columns to convert to logical}
}
\value{
data.frame
}
\description{
Split V(D)J meta.data columns into vectors
}
