% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_vdj}
\alias{filter_vdj}
\title{Subset Seurat object based on VDJ meta.data}
\usage{
filter_vdj(sobj_in, ..., cdr3_col = "cdr3s_aa")
}
\arguments{
\item{sobj_in}{Seurat object containing CDR3 sequences}

\item{...}{Expression to use for filtering object. To filter based on
receptor chains and CDR3 sequences use the terms \code{.chains} and \code{.seqs}.}

\item{cdr3_col}{meta.data column containing CDR3 sequences to use for
filtering}
}
\value{
Subsetted Seurat object
}
\description{
Subset Seurat object based on VDJ meta.data
}
