% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_vdj}
\alias{filter_vdj}
\title{Filter V(D)J meta.data}
\usage{
filter_vdj(
  sobj_in,
  filt,
  clonotype_col = "clonotype_id",
  filter_cells = FALSE,
  sep = ";",
  vdj_cols = NULL
)
}
\arguments{
\item{sobj_in}{Seurat object containing V(D)J data}

\item{filt}{Condition to use for filtering meta.data}

\item{clonotype_col}{meta.data column containing clonotype IDs. This column
is used to determine which cells have V(D)J data. If clonotype_col is set to
NULL, filtering is performed regardless of whether V(D)J data is present for
the cell}

\item{filter_cells}{Should cells be removed from object? If set FALSE
(default) V(D)J data will be removed from meta.data but no cells will be
removed from the object}

\item{sep}{Separator to use for expanding meta.data columns}

\item{vdj_cols}{meta.data columns containing VDJ data to use for filtering.
If set to NULL (recommended) columns are automatically selected.}
}
\value{
Seurat object
}
\description{
Filter V(D)J meta.data
}
