% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_vdj.R
\name{filter_vdj}
\alias{filter_vdj}
\title{Filter V(D)J data}
\usage{
filter_vdj(
  input,
  filt,
  clonotype_col = "cdr3_nt",
  filter_cells = FALSE,
  vdj_cols = NULL,
  sep = ";"
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{filt}{Condition to use for filtering meta.data}

\item{clonotype_col}{meta.data column containing clonotype IDs. This column
is used to determine which cells have V(D)J data. If clonotype_col is set to
NULL, filtering is performed regardless of whether V(D)J data is present for
the cell}

\item{filter_cells}{Should cells be removed from the object? If FALSE
(default) V(D)J data will be removed from the meta.data but no cells will be
removed from the object.}

\item{vdj_cols}{meta.data columns containing VDJ data to use for filtering.}

\item{sep}{Separator used for storing per cell V(D)J data
If set to NULL (recommended) columns are automatically selected based on the
given separator.}
}
\value{
Object with filtered meta.data
}
\description{
Filter V(D)J data
}
