% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_vdj}
\alias{filter_vdj}
\title{Subset Seurat object based on VDJ meta.data}
\usage{
filter_vdj(
  sobj_in,
  filt = NULL,
  new_col = NULL,
  true = TRUE,
  false = FALSE,
  clonotype_col = "clonotype_id",
  vdj_cols = c("chains", "cdr3"),
  sep = ";",
  return_seurat = TRUE
)
}
\arguments{
\item{sobj_in}{Seurat object containing VDJ data}

\item{filt}{Condition to use for filtering object. If set to NULL, all cells
will be returned. An argument must be passed to filt and/or new_col.}

\item{new_col}{Instead of filtering object create a new column with values
based on filtering condition. If filt is set to NULL, the values in new_col
will be based on the expression passed to true.}

\item{true}{Expression to generate values for new_col when filtering
condition evaluates to TRUE}

\item{false}{Expression to generate values for new_col when filtering
condition evaluates to FALSE}

\item{clonotype_col}{meta.data column containing clonotype IDs. This column
is used to determine which cells have VDJ data. If clonotype_col is set to
NULL, filtering is performed regardless of whether VDJ data is present for
the cell.}

\item{vdj_cols}{The names of meta.data columns to expand for filtering}

\item{sep}{Separator to use for expanding meta.data columns}

\item{return_seurat}{Return a Seurat object. If set to FALSE, the meta.data
table is returned.}
}
\value{
Filtered Seurat object
}
\description{
Subset Seurat object based on VDJ meta.data
}
