% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_outline.R
\name{geom_outline}
\alias{geom_outline}
\title{Highlight points}
\usage{
geom_outline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  geom = "point",
  outline_size = 1,
  outline_color = "black",
  outline_position = "all"
)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are often
aesthetics, used to set an aesthetic to a fixed value, like \code{colour = "red"}
or \code{size = 3}. They may also be parameters to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a
warning. If \code{TRUE}, missing values are silently removed.}

\item{outline_size}{Outline size}

\item{outline_color}{Outline color}

\item{outline_position}{Outline position}
}
\description{
Highlight points
}
