% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-vdj.R
\name{import_vdj}
\alias{import_vdj}
\title{Import V(D)J data}
\usage{
import_vdj(
  input = NULL,
  vdj_dir,
  prefix = "",
  cell_prefix = NULL,
  filter_contigs = TRUE,
  filter_paired = FALSE,
  define_clonotypes = NULL,
  sep = ";"
)
}
\arguments{
\item{input}{Object containing single cell data, if set to NULL a data.frame
containing V(D)J results will be returned}

\item{vdj_dir}{Directory containing the output from cellranger vdj. A vector
or named vector can be given to load data from several runs. If a named
vector is given, the cell barcodes will be prefixed with the provided names.
This mimics the behavior of the Read10X function found in the Seurat
package. Cell barcode prefixes can also be provided using the cell_prefix
argument.}

\item{prefix}{Prefix to add to new columns}

\item{cell_prefix}{Prefix to add to cell barcodes, this is helpful when
loading data from multiple runs into a single object. If set to NULL, cell
barcode prefixes will be automatically generated in a similar way as the
Read10X function found in the Seurat package.}

\item{filter_contigs}{Only include chains with at least one productive and
full length contig. It is highly recommended that filter_contigs is set to
TRUE when using the define_clonotypes argument.}

\item{filter_paired}{Only include clonotypes with paired chains. For TCR
data each clonotype must have at least one TRA and TRB chain, for BCR data
each clonotype must have at least one IGH chain and at least one IGK or IGL
chain.}

\item{define_clonotypes}{Define clonotype IDs based on V(D)J data. This is
useful if the V(D)J datasets being loaded do not have consistent clonotype
IDs, e.g., clonotype1 is not the same across samples.
'cdr3aa' will use the CDR3 amino acid sequence, 'cdr3nt' will use the CDR3
nucleotide sequence, and 'cdr3_gene' will use the combination of the CDR3
nucleotide sequence and the V(D)J genes. Set to NULL (default) to use the
clonotype IDs already present in the input data.}

\item{sep}{Separator to use for storing per cell V(D)J data}
}
\value{
Single cell object or data.frame with added V(D)J data
}
\description{
Import V(D)J data
}
