% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_vdj}
\alias{import_vdj}
\title{Add VDJ data to Seurat object}
\usage{
import_vdj(
  sobj_in,
  vdj_dir,
  prefix = "",
  cell_prefix = "",
  filter_contigs = TRUE,
  sep = ";"
)
}
\arguments{
\item{sobj_in}{Seurat object}

\item{vdj_dir}{cellranger vdj output directory}

\item{prefix}{Prefix to add to new meta.data columns}

\item{cell_prefix}{Prefix to add to cell barcodes}

\item{filter_contigs}{Only include chains with at least one productive
contig}

\item{sep}{Separator to use for storing per cell clonotype information in
the meta.data}
}
\value{
Seurat object with VDJ data added to meta.data
}
\description{
Add VDJ data to Seurat object
}
