% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_vdj}
\alias{import_vdj}
\title{Add VDJ data to Seurat object}
\usage{
import_vdj(sobj_in, vdj_dir, include_chains = NULL, prefix = "")
}
\arguments{
\item{sobj_in}{Seurat object}

\item{vdj_dir}{cellranger vdj output directory}

\item{include_chains}{Only inlude clonotypes that have the indicated chains}

\item{prefix}{Prefix to add to new meta.data columns}
}
\value{
Seurat object with VDJ data added to meta.data
}
\description{
Add VDJ data to Seurat object
}
