% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vdj.R
\name{import_vdj}
\alias{import_vdj}
\alias{import_vdj.default}
\alias{import_vdj.Seurat}
\title{Import V(D)J data}
\usage{
import_vdj(...)

\method{import_vdj}{default}(
  vdj_dir,
  prefix = "",
  cell_prefix = NULL,
  filter_contigs = TRUE,
  sep = ";",
  ...
)

\method{import_vdj}{Seurat}(
  input,
  vdj_dir,
  prefix = "",
  cell_prefix = NULL,
  filter_contigs = TRUE,
  sep = ";",
  ...
)
}
\arguments{
\item{...}{Arguments passed to other methods}

\item{vdj_dir}{Directory containing the output from cellranger vdj. A vector
or named vector can be given to load data from several runs. If a named
vector is given, the cell barcodes will be prefixed with the provided names.
This mimics the behavior of the Read10X function found in the Seurat
package. Cell barcode prefixes can also be provided using the cell_prefix
argument.}

\item{prefix}{Prefix to add to new columns}

\item{cell_prefix}{Prefix to add to cell barcodes, this is helpful when
loading data from multiple runs into a single object. If set to NULL, cell
barcode prefixes will be automatically generated in a similar way as the
Read10X function found in the Seurat package.}

\item{filter_contigs}{Only include chains with at least one productive
contig}

\item{sep}{Separator to use for storing per cell clonotype information}

\item{input}{Single cell object, if set to NULL a data.frame containing
V(D)J results will be returned}
}
\value{
Single cell object or data.frame with added V(D)J data
}
\description{
Import V(D)J data
}
