% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-vdj.R
\name{mutate_meta}
\alias{mutate_meta}
\title{Modify object meta.data}
\usage{
mutate_meta(input, fn, ...)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{fn}{Function or formula to use for modifying the meta.data. If a
formula is provided, use .x to refer to the meta.data table.}

\item{...}{Arguments to pass to the provided function}
}
\value{
Object with mutated meta.data
}
\description{
Modify object meta.data
}
