% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-vdj.R
\name{mutate_vdj}
\alias{mutate_vdj}
\title{Mutate V(D)J data}
\usage{
mutate_vdj(
  input,
  ...,
  clonotype_col = "clonotype_id",
  vdj_cols = NULL,
  return_df = FALSE,
  sep = ";"
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{...}{Name-value pairs to use for creating or modifying meta.data
columns}

\item{clonotype_col}{meta.data column containing clonotype IDs. This column
is used to determine which cells have V(D)J data.}

\item{vdj_cols}{meta.data columns containing V(D)J data to use for filtering.
If set to NULL (recommended) V(D)J is automatically selected by
identifying columns that have NAs in the same rows as clonotype_col.}

\item{return_df}{Return results as a data.frame. If set to FALSE, results
will be added to the input object.}

\item{sep}{Separator used for storing per cell V(D)J data}
}
\value{
Object with mutated meta.data
}
\description{
Mutate V(D)J data
}
