% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_vdj}
\alias{mutate_vdj}
\title{Mutate V(D)J meta.data}
\usage{
mutate_vdj(sobj_in, ..., clonotype_col = "cdr3_nt", sep = ";", vdj_cols = NULL)
}
\arguments{
\item{sobj_in}{Seurat object containing V(D)J data}

\item{...}{Name-value pairs to use for creating or modifying meta.data
columns}

\item{clonotype_col}{meta.data column containing clonotype IDs. This column
is used to determine which cells have V(D)J data}

\item{sep}{Separator to use for expanding meta.data columns}

\item{vdj_cols}{meta.data columns containing VDJ data to use for filtering.
If set to NULL (recommended) columns are automatically selected.}
}
\value{
Seurat object
}
\description{
Mutate V(D)J meta.data
}
