% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot clonotype abundance}
\usage{
plot_abundance(
  sobj_in,
  clonotype_col = "clonotype_id",
  cluster_col = NULL,
  abundance_col = NULL,
  plot_colors = NULL,
  plot_levels = NULL,
  ...
)
}
\arguments{
\item{sobj_in}{Seurat object containing VDJ data}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating clonotype abundance}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{abundance_col}{meta.data column containing pre-calculated abundances}

\item{plot_colors}{Character vector containing colors to use for plotting}

\item{plot_levels}{Character vector containing levels to use for ordering}

\item{...}{Additional arguments to pass to geom_line}
}
\value{
Seurat object with calculated clonotype abundance
}
\description{
Plot clonotype abundance
}
