% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-abundance.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot clonotype abundance}
\usage{
plot_abundance(
  input,
  cluster_col = NULL,
  clonotype_col = "clonotype_id",
  type = "bar",
  yaxis = "percent",
  plot_colors = NULL,
  plot_lvls = NULL,
  label_col = clonotype_col,
  n_clonotypes = 10,
  color_col = cluster_col,
  label_aes = list(),
  facet_rows = 1,
  facet_scales = "free_x",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance. Clonotypes will be
plotted separately for each cluster.}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating clonotype abundance}

\item{type}{Type of plot to create, can be 'bar' or 'line'}

\item{yaxis}{Units to plot on the y-axis, either 'frequency' or 'percent'}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{label_col}{meta.data column to use for labeling clonotypes. This is
useful if clonotype_col contains names that are too long to include on the
plot.}

\item{n_clonotypes}{Number of clonotypes to plot. If type is set to 'line',
this will specify the number of clonotypes to label.}

\item{color_col}{meta.data column to use for coloring clonotypes.}

\item{label_aes}{Named list providing additional aesthetics (color, size,
etc.) for clonotype labels when creating line graph}

\item{facet_rows}{The number of facet rows, use this when separate bar
graphs are created for each cell cluster}

\item{facet_scales}{This passes a scales specification to
ggplot2::facet_wrap, can be 'fixed', 'free', 'free_x', or 'free_y'. 'fixed'
will cause plot facets to share the same scales. Use this when separate bar
graphs are created for each cell cluster.}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Plot clonotype abundance
}
\examples{
# Plot clonotype abundance using all cells
plot_abundance(
  vdj_so,
  clonotype_col = "clonotype_id"
)

# Plot clonotype abundance separately for each cell cluster
plot_abundance(
  vdj_so,
  cluster_col = "orig.ident"
)

# Create line graph
plot_abundance(
  vdj_so,
  cluster_col = "orig.ident",
  type = "line"
)

# Plot the frequency of each clonotype instead of percentage
plot_abundance(
  vdj_so,
  cluster_col = "orig.ident",
  yaxis = "frequency"
)

# Specify colors to use for each cell cluster
plot_abundance(
  vdj_so,
  cluster_col = "orig.ident",
  plot_colors = c(avid_1 = "blue", avid_2 = "red")
)

# Specify order to use for plotting cell clusters
plot_abundance(
  vdj_so,
  cluster_col = "orig.ident",
  plot_lvls = c("avid_2", "avid_1")
)

# Specify meta.data column containing labels to use for each clonotype
plot_abundance(
  vdj_so,
  label_col = "cdr3"
)

# Specify the number of top clonotypes to plot
plot_abundance(
  vdj_so,
  n_clonotypes = 5
)

# Color bars based on meta.data column
plot_abundance(
  vdj_so,
  color_col = "orig.ident"
)

}
