% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot clonotype abundance}
\usage{
plot_abundance(
  input,
  cluster_col = NULL,
  clonotype_col = "clonotype_id",
  type = "bar",
  yaxis = "percent",
  plot_colors = NULL,
  plot_lvls = NULL,
  label_col = clonotype_col,
  n_clonotypes = 10,
  color_col = cluster_col,
  label_aes = list(),
  facet_rows = 1,
  facet_scales = "free_x",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating clonotype abundance}

\item{type}{Type of plot to create, can be 'bar' or 'line'}

\item{yaxis}{Units to plot on the y-axis, either 'frequency' or 'percent'}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{label_col}{meta.data column to use for labeling clonotypes. This is
useful if clonotype_col contains names that are too long to include on the
plot.}

\item{n_clonotypes}{Number of clonotypes to plot. If type is set to 'line',
this will specify the number of clonotypes to label.}

\item{color_col}{meta.data column to use for coloring bars}

\item{label_aes}{Named list providing additional label aesthetics (color,
size, etc.)}

\item{facet_rows}{The number of facet rows. Use this argument if type is set
to 'bar'}

\item{facet_scales}{If type is 'bar', this argument passes a scales
specification to facet_wrap, can be 'fixed', 'free', 'free_x', or 'free_y'}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Plot clonotype abundance
}
