% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cell_count}
\alias{plot_cell_count}
\title{Plot cell counts for clusters}
\usage{
plot_cell_count(
  sobj_in,
  x,
  fill_col = NULL,
  facet_col = NULL,
  yaxis = "fraction",
  plot_colors = NULL,
  plot_lvls = NULL,
  na_color = "grey90",
  n_label = TRUE,
  label_aes = list(),
  facet_rows = 1,
  facet_scales = "free_x",
  ...
)
}
\arguments{
\item{sobj_in}{Seurat object or data.frame containing data}

\item{x}{meta.data column containing clusters to use for creating bars}

\item{fill_col}{meta.data column to use for coloring bars}

\item{facet_col}{meta.data column containing groups to use for spliting plot
into facets}

\item{yaxis}{Units to plot on the y-axis, either "fraction" or "counts"}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{na_color}{Color to use for missing values}

\item{n_label}{Include label showing the number of cells represented by each
bar}

\item{label_aes}{Named list providing additional label aesthetics
(color, size, etc.)}

\item{facet_rows}{The number of facet rows. Use this argument if facet_col
is specified}

\item{facet_scales}{If facet_col is used, this argument passes a scales
specification to facet_wrap. Can be "fixed", "free", "free_x", or "free_y"}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Plot cell counts for clusters
}
