% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plot repertoire diversity}
\usage{
plot_diversity(
  input,
  cluster_col = NULL,
  method = abdiv::simpson,
  clonotype_col = "clonotype_id",
  plot_colors = NULL,
  plot_lvls = NULL,
  facet_rows = 1,
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{method}{Function to use for calculating diversity. A named list of
functions can be passed to plot multiple diversity metrics,
e.g. list(simpson = abdiv::simpson, shannon = abdiv::shannon)}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating clonotype abundance}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{facet_rows}{The number of facet rows, use this argument if a list of
functions is passed to method}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Plot repertoire diversity
}
