% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-diversity.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plot repertoire diversity}
\usage{
plot_diversity(
  input,
  data_col = "clonotype_id",
  cluster_col = NULL,
  group_col = NULL,
  method = abdiv::simpson,
  downsample = FALSE,
  n_boots = 1,
  chain = NULL,
  chain_col = "chains",
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  facet_rows = 1,
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing clonotype IDs to use for
calculating repertoire diversity}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{group_col}{meta.data column to use for grouping IDs present in
cluster_col}

\item{method}{Function to use for calculating diversity. A named list of
functions can be passed to plot multiple diversity metrics,
e.g. list(simpson = abdiv::simpson, shannon = abdiv::shannon)}

\item{downsample}{Downsample clusters to the same size when calculating
diversity metrics}

\item{n_boots}{Number of bootstrap replicates}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{facet_rows}{The number of facet rows for final plot, use this argument
if a list of functions is passed to method}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Plot repertoire diversity
}
\examples{
# Plot diversity using all cells
plot_diversity(vdj_so)

# Specify method to use for calculating repertoire diversity
plot_diversity(
  vdj_sce,
  method = abdiv::shannon
)

# Plot diversity separately for each cell cluster
plot_diversity(
  vdj_so,
  cluster_col = "orig.ident"
)

# Plot multiple diversity metrics
plot_diversity(
  vdj_sce,
  cluster_col = "orig.ident",
  method = list(simpson = abdiv::simpson, shannon = abdiv::shannon)
)

# Specify colors to use for each cell cluster
plot_diversity(
  vdj_so,
  cluster_col = "orig.ident",
  plot_colors = c(avid_2 = "green", avid_1 = "purple")
)

# Specify order to use for plotting cell clusters
plot_diversity(
  vdj_sce,
  cluster_col = "orig.ident",
  plot_lvls = c("avid_2", "avid_1")
)

# Specify how to organize facets when plotting multiple metrics

mets <- list(
  simpson = abdiv::simpson,
  shannon = abdiv::shannon
)

plot_diversity(
  vdj_so,
  method = mets,
  facet_rows = 2
)

}
