% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plot repertoire diversity}
\usage{
plot_diversity(
  input,
  clonotype_col = "cdr3_nt",
  cluster_col = NULL,
  method = abdiv::simpson,
  plot_colors = NULL,
  plot_lvls = NULL,
  facet_rows = 1,
  ...
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating clonotype abundance}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{method}{Method to use for calculating diversity. Can also pass a named
list to use multiple methods.}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{facet_rows}{The number of facet rows, use this argument if a list of
functions is passed to method}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Plot repertoire diversity
}
