% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-diversity.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plot repertoire diversity}
\usage{
plot_diversity(
  input,
  data_col,
  cluster_col = NULL,
  group_col = NULL,
  method = abdiv::simpson,
  downsample = FALSE,
  n_boots = 0,
  chain = NULL,
  chain_col = "chains",
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  panel_nrow = NULL,
  panel_scales = "free",
  n_label = TRUE,
  label_params = list(),
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing values to use for
calculating diversity, e.g. 'clonotype_id'}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{group_col}{meta.data column to use for grouping clusters present in
cluster_col}

\item{method}{Function to use for calculating diversity, e.g.
abdiv::simpson. A named list of functions can be passed to plot multiple
diversity metrics, e.g.
list(simpson = abdiv::simpson, shannon = abdiv::shannon)}

\item{downsample}{Downsample clusters to the same size when calculating
diversity metrics}

\item{n_boots}{Number of bootstrap replicates for calculating standard error,
if n_boots is 0 this will be skipped.}

\item{chain}{Chain to use for calculating diversity. To calculate diversity
for a single chain, the column passed to the data_col argument must contain
per-chain data such as CDR3 sequences. Set to NULL to include all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{panel_nrow}{The number of rows to use for arranging plot panels}

\item{panel_scales}{Should scales for plot panels be fixed or free. This
passes a scales specification to ggplot2::facet_wrap, can be 'fixed', 'free',
'free_x', or 'free_y'. 'fixed' will cause panels to share the same scales.}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Plot repertoire diversity
}
\examples{
# Plot diversity using all cells
plot_diversity(
  vdj_so,
  data_col = "clonotype_id"
)

# Specify method to use for calculating repertoire diversity
plot_diversity(
  vdj_sce,
  data_col = "clonotype_id",
  method   = abdiv::shannon
)

# Plot diversity separately for each cell cluster
plot_diversity(
  vdj_so,
  data_col    = "clonotype_id",
  cluster_col = "orig.ident"
)

# Plot multiple diversity metrics
mets <- list(
  simpson = abdiv::simpson,
  shannon = abdiv::shannon
)

plot_diversity(
  vdj_sce,
  data_col    = "clonotype_id",
  cluster_col = "orig.ident",
  method      = mets
)

# Specify colors to use for each cell cluster
plot_diversity(
  vdj_so,
  data_col    = "clonotype_id",
  cluster_col = "orig.ident",
  plot_colors = c(avid_2 = "green", avid_1 = "purple")
)

# Specify order to use for plotting cell clusters
plot_diversity(
  vdj_sce,
  data_col    = "clonotype_id",
  cluster_col = "orig.ident",
  plot_lvls   = c("avid_2", "avid_1")
)

# Specify how to organize panels when plotting multiple metrics
plot_diversity(
  vdj_so,
  data_col   = "clonotype_id",
  method     = mets,
  panel_nrow = 2
)

}
\seealso{
\code{\link[=calc_diversity]{calc_diversity()}}, \code{\link[=plot_rarefaction]{plot_rarefaction()}}
}
