% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_features}
\alias{plot_features}
\alias{plot_features.default}
\alias{plot_features.Seurat}
\alias{plot_vdj_umap}
\alias{plot_vdj_umap.default}
\alias{plot_vdj_umap.Seurat}
\title{Create 2D feature plot}
\usage{
plot_features(input, ...)

\method{plot_features}{default}(
  input,
  feature,
  x = "UMAP_1",
  y = "UMAP_2",
  plot_colors = NULL,
  plot_lvls = NULL,
  min_q = NULL,
  max_q = NULL,
  na_color = "grey90",
  ...
)

\method{plot_features}{Seurat}(
  input,
  feature,
  x = "UMAP_1",
  y = "UMAP_2",
  data_slot = "data",
  plot_colors = NULL,
  plot_lvls = NULL,
  min_q = NULL,
  max_q = NULL,
  na_color = "grey90",
  ...
)

plot_vdj_umap(input, ...)

\method{plot_vdj_umap}{default}(
  input,
  data_col,
  x = "UMAP_1",
  y = "UMAP_2",
  summary_fn = NULL,
  chain = NULL,
  plot_lvls = NULL,
  min_q = NULL,
  max_q = NULL,
  na_color = "grey90",
  chain_col = "chains",
  sep = ";",
  ...
)

\method{plot_vdj_umap}{Seurat}(
  input,
  data_col,
  x = "UMAP_1",
  y = "UMAP_2",
  data_slot = "data",
  summary_fn = NULL,
  chain = NULL,
  plot_lvls = NULL,
  min_q = NULL,
  max_q = NULL,
  na_color = "grey90",
  chain_col = "chains",
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{...}{Additional arguments to pass to ggplot geom}

\item{feature}{Variable to use for coloring points}

\item{x}{Variable to plot on x-axis}

\item{y}{Variable to plot on y-axis}

\item{plot_colors}{Vector of colors to use for plot}

\item{plot_lvls}{Levels to use for ordering feature}

\item{min_q}{Minimum quantile cutoff for color scale.}

\item{max_q}{Maximum quantile cutoff for color scale.}

\item{na_color}{Color to use for missing values}

\item{data_slot}{Slot in the Seurat object to pull data}

\item{data_col}{meta.data column containing V(D)J data to use for coloring
cells}

\item{summary_fn}{Function to use for summarizing values for each cell, this
can be either a function, e.g. mean, or a purrr-style lambda,
e.g. ~ mean(.x, na.rm = TRUE). If NULL, the default for numeric values is
mean and the default for other types is ~ paste0(.x, collapse = sep). This
is necessary since some V(D)J metrics will have multiple values for each
cell.}

\item{chain}{Chain(s) to use for plotting, set to NULL to include all chains}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}
}
\value{
ggplot object
}
\description{
Create 2D feature plot
}
