% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_features}
\alias{plot_features}
\title{Create two dimensional scatter plot}
\usage{
plot_features(
  obj_in,
  x = "UMAP_1",
  y = "UMAP_2",
  feature,
  data_slot = "data",
  pt_size = 0.25,
  plot_colors = NULL,
  feat_levels = NULL,
  split_id = NULL,
  split_levels = NULL,
  min_pct = NULL,
  max_pct = NULL,
  lm_line = F,
  cor_label = c(0.8, 0.9),
  label_size = 3.7,
  ...
)
}
\arguments{
\item{obj_in}{Seurat object or data.frame containing data for plotting}

\item{x}{Variable to plot on x-axis}

\item{y}{Variable to plot on y-axis}

\item{feature}{Variable to use for coloring points}

\item{data_slot}{Slot in the Seurat object to pull data from}

\item{pt_size}{Point size}

\item{plot_colors}{Vector of colors to use for plot}

\item{feat_levels}{Levels to use for ordering feature}

\item{split_id}{Variable to use for splitting plot into facets}

\item{split_levels}{Levels to use for ordering plot facets}

\item{min_pct}{Minimum value to plot for feature}

\item{max_pct}{Maximum value to plot for feature}

\item{lm_line}{Add regression line to plot}

\item{cor_label}{Position of correlation coefficient label}

\item{label_size}{Size of correlation coefficient label}

\item{...}{Additional parameters to pass to facet_wrap}
}
\value{
ggplot object
}
\description{
Create two dimensional scatter plot
}
