% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vdj.R
\name{plot_features}
\alias{plot_features}
\alias{plot_features.default}
\alias{plot_features.Seurat}
\alias{plot_vdj_umap}
\alias{plot_vdj_umap.default}
\alias{plot_vdj_umap.Seurat}
\title{Create 2D feature plot}
\usage{
plot_features(input, ...)

\method{plot_features}{default}(
  input,
  feature,
  x = "UMAP_1",
  y = "UMAP_2",
  plot_colors = NULL,
  plot_lvls = NULL,
  min_q = NULL,
  max_q = NULL,
  na_color = "grey90",
  ...
)

\method{plot_features}{Seurat}(
  input,
  feature,
  x = "UMAP_1",
  y = "UMAP_2",
  data_slot = "data",
  plot_colors = NULL,
  plot_lvls = NULL,
  min_q = NULL,
  max_q = NULL,
  na_color = "grey90",
  ...
)

plot_vdj_umap(input, ...)

\method{plot_vdj_umap}{default}(
  input,
  data_col,
  x = "UMAP_1",
  y = "UMAP_2",
  summary_fn = NULL,
  chain = NULL,
  plot_lvls = NULL,
  min_q = NULL,
  max_q = NULL,
  na_color = "grey90",
  chain_col = "chains",
  sep = ";",
  ...
)

\method{plot_vdj_umap}{Seurat}(
  input,
  data_col,
  x = "UMAP_1",
  y = "UMAP_2",
  data_slot = "data",
  summary_fn = NULL,
  chain = NULL,
  plot_lvls = NULL,
  min_q = NULL,
  max_q = NULL,
  na_color = "grey90",
  chain_col = "chains",
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}

\item{feature}{Variable to use for coloring points}

\item{x}{Variable to plot on x-axis}

\item{y}{Variable to plot on y-axis}

\item{plot_colors}{Vector of colors to use for plotting}

\item{plot_lvls}{Levels to use for ordering feature}

\item{min_q}{Minimum quantile cutoff for color scale.}

\item{max_q}{Maximum quantile cutoff for color scale.}

\item{na_color}{Color to use for missing values}

\item{data_slot}{Slot in the Seurat object to pull data}

\item{data_col}{meta.data column containing V(D)J data to use for coloring
cells}

\item{summary_fn}{Function to use for summarizing values for each cell, possible
values can be either a function, e.g. mean, or a purrr-style lambda,
e.g. ~ mean(.x, na.rm = TRUE). If NULL, the mean will be calculated for
numeric values, non-numeric columns will be combined into a single string.}

\item{chain}{Chain(s) to use for filtering data before plotting. If NULL
data will not be filtered based on chain.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}
}
\value{
ggplot object
}
\description{
Create a scatter plot with cells colored based on the provided feature.
This can be used to create dimensional reduction plots
(e.g. UMAP, tSNE, PCA) or to compare different cell attributes
(e.g. CD4 vs CD8 expression).

plot_vdj_umap() allows per-chain V(D)J data to be summarized
and plotted for each cell.
}
\examples{
# Create UMAP with cells colored based on sample name
plot_features(
  tiny_so,
  feature = "orig.ident"
)

# Compare UMI counts for each cell with number of genes detected
plot_features(
  tiny_sce,
  feature = "orig.ident",
  x = "nFeature_RNA",
  y = "nCount_RNA"
)

# Plot average CDR3 length for each cell for light chains
plot_vdj_umap(
  vdj_so,
  data_col = "cdr3_length",
  summary_fn = mean,
  chain = c("IGK", "IGL")
)

# Plot median number of insertions for each cell
plot_vdj_umap(
  vdj_so,
  data_col = "n_insertion",
  summary_fn = stats::median
)

}
