% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-abundance.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot frequency of cell groups/labels present in the object}
\usage{
plot_frequency(
  input,
  data_col,
  cluster_col = NULL,
  group_col = NULL,
  yaxis = "percent",
  plot_colors = NULL,
  plot_lvls = NULL,
  facet_rows = NULL,
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing cell labels to use for
calculating abundance. To calculate clonotype abundance, provide the column
containing clonotype IDs, to calculate isotype abundance provide the column
containing cell isotypes. By default clonotype_id is used for calculations.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance. Clonotypes will be
plotted separately for each cluster.}

\item{group_col}{meta.data column to use for grouping cluster IDs present in
cluster_col}

\item{yaxis}{Units to plot on the y-axis, either 'frequency' or 'percent'}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{facet_rows}{The number of facet rows, use this when separate bar
graphs are created for each cell cluster}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Plot the frequency of each cell label present in the provided meta.data
column. This is useful for comparing the proportion of cells belonging to
different samples, cell types, isotypes, etc. To compare clonotype
frequency, use the plot_clonal_abundance() function.
}
