% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-frequency.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot frequency of cell groups present in object}
\usage{
plot_frequency(
  input,
  data_col,
  cluster_col = NULL,
  group_col = NULL,
  units = "percent",
  stack = TRUE,
  plot_colors = NULL,
  plot_lvls = NULL,
  trans = "identity",
  n_top = NULL,
  other_label = "other",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing cell labels to use for
calculating frequency, e.g. cell types, isotypes etc.}

\item{cluster_col}{meta.data column containing cluster IDs (or patients,
treatment conditions, etc.) to use when calculating frequency. Calculations
will be performed separately for each cluster.}

\item{group_col}{meta.data column to use for grouping cluster IDs present in
cluster_col. This is useful when there are multiple replicates or patients
for each treatment condition.}

\item{units}{Units to plot on the y-axis, either 'frequency' or 'percent'}

\item{stack}{If TRUE, stacked bargraphs will be generated.}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Levels to use for ordering clusters or groups}

\item{trans}{Transformation to use for plotting data, e.g. 'log10'. By
default values are not transformed, refer to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}
for more options. Values can only be transformed when stack is \code{FALSE}}

\item{n_top}{Number of top cell labels present in data_col to show on plot,
other cells will be labeled based on the other_label argument. If \code{NULL},
this will be automatically selected.}

\item{other_label}{Label to use for 'other' cells, if \code{NULL} all cell labels
present in data_col will be displayed on the plot.}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Plot the frequency of each cell label present in the provided meta.data
column. This is useful for comparing the proportion of cells belonging to
different samples, cell types, clonotypes, isotypes, etc.
}
\examples{
# Plot frequency of different isotypes
plot_frequency(
  vdj_so,
  data_col = "isotype"
)

# Plot frequency separately for cell clusters
plot_frequency(
  vdj_sce,
  data_col    = "isotype",
  cluster_col = "orig.ident"
)

# Create grouped bar graphs
plot_frequency(
  vdj_sce,
  data_col    = "isotype",
  cluster_col = "orig.ident",
  stack       = FALSE
)

# Plot number of cells on the y-axis
plot_frequency(
  vdj_so,
  data_col    = "seurat_clusters",
  cluster_col = "orig.ident",
  units       = "frequency"
)

}
\seealso{
\code{\link[=calc_frequency]{calc_frequency()}}, \code{\link[=plot_clone_frequency]{plot_clone_frequency()}}
}
