% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-gene-usage.R
\name{plot_gene_pairs}
\alias{plot_gene_pairs}
\title{Plot usage of V(D)J gene pairs}
\usage{
plot_gene_pairs(
  input,
  data_col,
  chains,
  cluster_col = NULL,
  group_col = NULL,
  genes = 20,
  method = "heatmap",
  plot_colors = NULL,
  plot_lvls = NULL,
  chain_col = global$chain_col,
  trans = "identity",
  units = "percent",
  rotate_labels = FALSE,
  panel_nrow = NULL,
  show_points = TRUE,
  n_label = NULL,
  label_params = list(),
  return_list = FALSE,
  sep = global$sep,
  ...
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing V(D)J genes identified for each
clonotype.}

\item{chains}{Chains to use for calculating usage of different gene pairs.
This should be a character vector containing the two chains to use for
calculations, e.g. \code{c("IGH", "IGK")}.}

\item{cluster_col}{meta.data column containing cell clusters to use when
calculating gene usage}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}
}
\value{
data.frame containing gene pair summary
}
\description{
Plot usage of V(D)J gene pairs
}
\seealso{
\code{\link[=calc_gene_usage(), calc_gene_pairs(), plot_gene_usage]{calc_gene_usage(), calc_gene_pairs(), plot_gene_usage()}}
}
