% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-seqs.R
\name{plot_motifs}
\alias{plot_motifs}
\title{Plot CDR3 sequence motifs}
\usage{
plot_motifs(
  input,
  data_col = global$cdr3_col,
  cluster_col = NULL,
  chain,
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  chain_col = global$chain_col,
  width = 0.75,
  align_end = "5",
  panel_nrow = NULL,
  panel_scales = "free",
  n_label = TRUE,
  label_params = list(),
  quiet = FALSE,
  sep = global$sep,
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing sequences to use for plotting.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells.}

\item{chain}{Chain to use for plotting sequences. Cells with more than one
of the provided chain will be excluded from the analysis.}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{chain_col}{meta.data column containing chains for each cell.}

\item{width}{Integer specifying how many residues to include, sequences
longer than width will be get trimmed based on the align_end argument,
sequences shorter than width will get removed. If a fraction is provided,
the width cutoff is set based on percent rank, i.e. a value of 0.75 would
select a width where at least 75\% of sequences are longer than the cutoff.}

\item{align_end}{End to use for aligning sequences, specify '5' or '3' to
align sequences at the 5' or 3' end when plotting.}

\item{panel_nrow}{The number of rows to use for arranging plot panels}

\item{panel_scales}{Should scales for plot panels be fixed or free. This
passes a scales specification to \code{ggplot2::facet_wrap()}, can be 'fixed', 'free',
'free_x', or 'free_y'. 'fixed' will cause panels to share the same scales.
Use this when separate bar graphs are created for each cell cluster.}

\item{n_label}{Location on plot where n label should be added, this can be
one of the following:
\itemize{
\item 'corner', display the total number of cells plotted in the top right
corner, the position of the label can be modified by passing \code{x} and \code{y}
specifications with the \code{label_params} argument
\item 'none', do not display the number of cells plotted
}}

\item{label_params}{Named list providing additional parameters to modify
n label aesthetics, e.g. list(size = 4, color = "red")}

\item{quiet}{If \code{TRUE} messages will not be displayed}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{...}{Additional parameters to pass to \code{\link[ggseqlogo:geom_logo]{ggseqlogo::geom_logo()}}}
}
\value{
ggplot object
}
\description{
Plot CDR3 sequence motifs
}
\examples{
# Cluster cells based on CDR3 amino acid sequences and plot sequence motifs
res <- cluster_sequences(
  vdj_so,
  data_col = "cdr3"
)

plot_motifs(
  res,
  data_col    = "cdr3",
  cluster_col = "cdr3_cluster_0.5",
  chain       = "IGK"
)

}
\seealso{
\code{\link[=cluster_sequences]{cluster_sequences()}}
}
