% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-diversity.R
\name{plot_rarefaction}
\alias{plot_rarefaction}
\title{Plot rarefaction curve}
\usage{
plot_rarefaction(
  input,
  data_col,
  cluster_col = NULL,
  method = "richness",
  n_boots = 50,
  chain = NULL,
  chain_col = "chains",
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  panel_nrow = NULL,
  panel_scales = "free",
  ci_alpha = 0.15,
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing values to use for
calculating diversity}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{method}{Method to use for calculating diversity}

\item{n_boots}{Number of bootstrap replicates for calculating standard error,
if n_boots is 0 this will be skipped.}

\item{chain}{Chain to use for calculating diversity. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{panel_nrow}{The number of rows to use for arranging plot panels}

\item{panel_scales}{Should scales for plot panels be fixed or free. This
passes a scales specification to ggplot2::facet_wrap, can be 'fixed', 'free',
'free_x', or 'free_y'. 'fixed' will cause panels to share the same scales.}

\item{ci_alpha}{Transparency to use when plotting 95\% confidence interval}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Plot rarefaction curve
}
\seealso{
\code{\link[=calc_diversity]{calc_diversity()}}, \code{\link[=plot_diversity]{plot_diversity()}}
}
