% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-diversity.R
\name{plot_rarefaction}
\alias{plot_rarefaction}
\title{Create rarefaction curve}
\usage{
plot_rarefaction(
  input,
  data_col,
  cluster_col = NULL,
  method = "richness",
  n_boots = 50,
  chain = NULL,
  chain_col = "chains",
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  facet_rows = 1,
  facet_scales = "free",
  ci_alpha = 0.15,
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing values to use for
calculating diversity}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells when calculating clonotype abundance}

\item{method}{Method to use for calculating diversity}

\item{n_boots}{Number of bootstrap replicates to use for calculating
standard error}

\item{chain}{Chain to use for calculating diversity. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Character vector containing levels for ordering}

\item{facet_rows}{The number of facet rows for final plot}

\item{facet_scales}{This passes a scales specification to
ggplot2::facet_wrap, can be 'fixed', 'free', 'free_x', or 'free_y'. 'fixed'
will cause plot facets to share the same scales.}

\item{ci_alpha}{Transparency to use when plotting 95\% confidence interval}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Create rarefaction curve
}
