% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vdj.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Create 2D scatter plot}
\usage{
plot_scatter(
  input,
  data_col = NULL,
  x = "UMAP_1",
  y = "UMAP_2",
  group_col = NULL,
  data_slot = "data",
  top = NULL,
  other_label = "other",
  plot_colors = NULL,
  plot_lvls = NULL,
  trans = "identity",
  panel_nrow = NULL,
  panel_scales = "fixed",
  min_q = NULL,
  max_q = NULL,
  na_color = "grey80",
  n_label = NULL,
  label_params = list(),
  ...,
  chain = NULL,
  chain_col = global$chain_col,
  summary_fn = NULL,
  sep = global$sep
)
}
\arguments{
\item{input}{Single cell object or data.frame, if a
data.frame is provided, cell barcodes should be stored as row names.}

\item{data_col}{Name of meta.data column or other variable (e.g. gene name)
to use for coloring points}

\item{x, y}{Name of meta.data column or other variable to plot on x and
y-axis}

\item{group_col}{meta.data column to use for splitting plot into panels}

\item{data_slot}{Slot to pull data from when \code{input} is a Seurat object}

\item{top}{To only show the top cell groups, provide
one of the following, all other cells will be labeled using the value
provided to the \code{other_label} argument. If \code{NULL} this will be automatically
set.
\itemize{
\item Integer specifying the number of top groups to show
\item Vector specifying the names of cell groups to show
}}

\item{other_label}{Label to use for 'other' cells when \code{top} is specified, if
\code{NULL} all cell groups will be shown.
\subsection{Aesthetics}{
}}

\item{plot_colors}{Character vector specifying colors to use for cell
clusters specified by cluster_col.}

\item{plot_lvls}{Character vector containing order to use for plotting cell
clusters specified by cluster_col.}

\item{trans}{Transformation to use when plotting data, e.g. 'log10'. By
default values are not transformed, refer to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}
for more options.}

\item{panel_nrow}{The number of rows to use for arranging plot panels}

\item{panel_scales}{Should scales for plot panels be fixed or free. This
passes a scales specification to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, can be 'fixed',
'free', 'free_x', or 'free_y'. 'fixed' will cause panels to share the same
scales.}

\item{min_q}{Minimum quantile cutoff for color scale.}

\item{max_q}{Maximum quantile cutoff for color scale.}

\item{na_color}{Color to use for missing values. If plotting V(D)J data,
cells lacking data will be plotted as \code{NA}s.}

\item{n_label}{Location on plot where n label should be added, this can be
any combination of the following:
\itemize{
\item 'corner', display the total number of cells plotted in the top right
corner, the position of the label can be modified by passing \code{x} and \code{y}
specifications with the \code{label_params} argument
\item 'legend', display the number of cells plotted for each group shown in the
plot legend
\item 'none', do not display the number of cells plotted
}}

\item{label_params}{Named list providing additional parameters to modify
n label aesthetics, e.g. list(size = 4, color = "red")}

\item{...}{Additional arguments to pass to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}, e.g.
color, size, etc.
\subsection{VDJ arguments}{
}}

\item{chain}{Chain(s) to use for filtering data before plotting. If NULL
data will not be filtered based on chain.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{summary_fn}{Function to use for summarizing per-chain values for each
cell, can be either a function, e.g. \code{mean}, or a purrr-style
lambda, e.g. \code{~ mean(.x, na.rm = TRUE)} where \code{.x} refers to the column. If
\code{NULL}, the mean will be calculated for numeric values, non-numeric columns
will be combined into a single string.}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}
}
\value{
ggplot object
}
\description{
Create a scatter plot with cells colored based on the provided feature.
This can be used to create dimensional reduction plots (e.g. UMAP, tSNE, PCA)
or to compare different cell attributes (e.g. CD4 vs CD8 expression).
When plotting V(D)J data, per-chain values will be summarized for each cell.
}
\seealso{
\code{\link[=summarize_vdj]{summarize_vdj()}} for more examples on how per-chain data can be
summarized for each cell
}
