% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Plot repertoire overlap}
\usage{
plot_similarity(
  input,
  clonotype_col = "cdr3_nt",
  cluster_col,
  method = abdiv::jaccard,
  plot_colors = NULL,
  ...
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{method}{Method to use for calculating similarity between clusters}

\item{plot_colors}{Character vector containing colors for plotting}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Plot repertoire overlap
}
