% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Plot repertoire overlap}
\usage{
plot_similarity(
  input,
  cluster_col,
  method = abdiv::jaccard,
  clonotype_col = "clonotype_id",
  plot_colors = NULL,
  plot_lvls = NULL,
  include_upper_triangle = TRUE,
  include_diagonal = TRUE,
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{method}{Method to use for calculating similarity between clusters}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Levels to use for ordering clusters}

\item{include_upper_triangle}{If FALSE, upper triangle will not be shown on
plot and rows/columns will not be clustered.}

\item{include_diagonal}{If FALSE, diagonal will not be shown on heatmap and
rows/columns will not be clustered.}

\item{...}{Additional arguments to pass to ComplexHeatmap::Heatmap()}
}
\value{
ggplot object
}
\description{
Plot repertoire overlap
}
\examples{
# Plot repertoire overlap
plot_similarity(
  vdj_so,
  cluster_col = "orig.ident"
)

# Specify method to use for calculating repertoire overlap
plot_similarity(
  vdj_sce,
  cluster_col = "orig.ident",
  method = abdiv::jaccard
)

# Specify colors to use for heatmap
plot_similarity(
  vdj_so,
  cluster_col = "orig.ident",
  plot_color = c("white", "red")
)

# Pass additional aesthetic parameters to ggplot2
plot_similarity(
  vdj_sce,
  cluster_col = "orig.ident",
  color = "black",
  size = 2
)

}
