% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Plot repertoire overlap}
\usage{
plot_similarity(
  input,
  data_col,
  cluster_col,
  group_col = NULL,
  method = abdiv::jaccard,
  chain = NULL,
  chain_col = "chains",
  plot_colors = NULL,
  plot_lvls = names(plot_colors),
  cluster_heatmap = TRUE,
  remove_upper_triangle = FALSE,
  remove_diagonal = remove_upper_triangle,
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{data_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{group_col}{meta.data column to use for grouping cluster IDs present in
cluster_col. This is useful when there are multiple replicates or patients
for each treatment condition.}

\item{method}{Method to use for comparing clusters, possible values are:
\itemize{
\item A function that takes two numeric vectors containing counts for each
clonotype in the object, such as the beta diversity functions provided by
the abdiv package. This will generate a heatmap.
\item 'count', count the number of clonotypes overlapping between each cluster,
this will generate a heatmap.
\item 'mds', perform multidimensional scaling, this will generate a scatter plot
\item 'circos', create circos plot implemented with circlize::chordDiagram()
}}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{chain_col}{meta.data column containing chains for each cell}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_lvls}{Levels to use for ordering clusters}

\item{cluster_heatmap}{If FALSE, rows and columns of heatmap will not be
clustered.}

\item{remove_upper_triangle}{If TRUE, upper triangle for heatmap will not
be shown.}

\item{remove_diagonal}{If TRUE, diagonal for heatmap will not be shown.}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to plotting function}
}
\description{
Plot repertoire overlap
}
\examples{
# Plot repertoire overlap
plot_similarity(
  vdj_so,
  cluster_col = "orig.ident"
)

# Specify method to use for calculating repertoire overlap
plot_similarity(
  vdj_sce,
  cluster_col = "orig.ident",
  method = abdiv::jaccard
)

# Specify colors to use for heatmap
plot_similarity(
  vdj_so,
  cluster_col = "orig.ident",
  plot_color = c("white", "red")
)

# Pass additional aesthetic parameters to ggplot2
plot_similarity(
  vdj_sce,
  cluster_col = "orig.ident",
  color = "black",
  size = 2
)

}
\seealso{
\code{\link[=calc_similarity]{calc_similarity()}}, \code{\link[circlize:chordDiagram]{circlize::chordDiagram()}}, \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}},
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}
}
