% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Plot repertoire overlap}
\usage{
plot_similarity(
  sobj_in,
  clonotype_col = NULL,
  cluster_col = NULL,
  method = "jaccard",
  plot_colors = NULL,
  ...
)
}
\arguments{
\item{sobj_in}{Seurat object containing V(D)J data or matrix containing
similarity values to plot}

\item{clonotype_col}{meta.data column containing clonotype IDs to use for
calculating overlap}

\item{cluster_col}{meta.data column containing cluster IDs to use for
calculating overlap}

\item{method}{Method to use for calculating similarity between clusters}

\item{plot_colors}{Character vector containing colors for plotting}

\item{...}{Additional arguments to pass to geom_tile}
}
\value{
ggplot object
}
\description{
Plot repertoire overlap
}
