% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_usage}
\alias{plot_usage}
\title{Plot V(D)J gene usage}
\usage{
plot_usage(
  sobj_in,
  gene_col,
  cluster_col = NULL,
  chain = NULL,
  plot_colors = NULL,
  plot_genes = NULL,
  n_genes = NULL,
  clust_levels = NULL,
  yaxis = "percent",
  ...,
  chain_col = "chains",
  sep = ";"
)
}
\arguments{
\item{sobj_in}{Seurat object containing V(D)J data}

\item{gene_col}{meta.data column containing genes used for each clonotype}

\item{cluster_col}{meta.data column containing cell clusters to use for
calculating gene usage}

\item{chain}{Chain to use for calculating gene usage. Set to NULL to include
all chains.}

\item{plot_colors}{Character vector containing colors to use for plotting}

\item{plot_genes}{Character vector of genes to plot}

\item{n_genes}{Number of top genes to plot based on average usage}

\item{clust_levels}{Levels to use for ordering clusters}

\item{yaxis}{Units to plot on the y-axis, either "frequency" or "percent"}

\item{...}{Additional arguments to pass to geom_tile}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator to use for expanding gene_col}
}
\value{
ggplot object
}
\description{
Plot V(D)J gene usage
}
