% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_usage}
\alias{plot_usage}
\title{Plot V(D)J gene usage}
\usage{
plot_usage(
  input,
  gene_cols,
  cluster_col = NULL,
  chain = NULL,
  type = "bar",
  plot_colors = NULL,
  plot_genes = NULL,
  n_genes = 50,
  plot_lvls = NULL,
  yaxis = "percent",
  chain_col = "chains",
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Object containing V(D)J data. If a data.frame is provided, the
cell barcodes should be stored as row names.}

\item{gene_cols}{meta.data column containing genes for each clonotype,
provide a vector with two column names to plot paired usage of genes}

\item{cluster_col}{meta.data column containing cell clusters to use for
calculating gene usage}

\item{chain}{Chain to use for calculating gene usage, set to NULL to include
all chains}

\item{type}{Type of plot to create, can be either 'heatmap' or 'bar', a
bar plot can only be created when a single gene_col is provided}

\item{plot_colors}{Character vector containing colors for plotting}

\item{plot_genes}{Character vector containing genes to plot}

\item{n_genes}{Number of top genes to plot based on usage. If cluster_col is
provided, top genes will be identified for each cluster.}

\item{plot_lvls}{Levels to use for ordering clusters}

\item{yaxis}{Units to plot on the y-axis, either 'frequency' or 'percent'}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Plot V(D)J gene usage
}
