% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vdj.R
\name{plot_vdj}
\alias{plot_vdj}
\alias{plot_vdj_reads}
\alias{plot_cdr3_length}
\alias{plot_vdj_indels}
\title{Plot continuous V(D)J data}
\usage{
plot_vdj(
  input,
  data_cols,
  per_cell = FALSE,
  summary_fn = mean,
  cluster_col = NULL,
  chain = NULL,
  type = "histogram",
  yaxis = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  alpha = 0.5,
  log_trans = FALSE,
  chain_col = "chains",
  sep = ";",
  ...
)

plot_vdj_reads(
  input,
  data_cols = c("reads", "umis"),
  cluster_col = NULL,
  chain = NULL,
  type = "violin",
  yaxis = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  log_trans = TRUE,
  chain_col = "chains",
  sep = ";",
  ...
)

plot_cdr3_length(
  input,
  data_cols = "cdr3_length",
  cluster_col = NULL,
  chain = NULL,
  type = "histogram",
  yaxis = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  log_trans = TRUE,
  chain_col = "chains",
  sep = ";",
  ...
)

plot_vdj_indels(
  input,
  data_cols = c("n_insertion", "n_deletion", "n_mismatch"),
  cluster_col = NULL,
  chain = NULL,
  type = "violin",
  yaxis = "frequency",
  plot_colors = NULL,
  plot_lvls = NULL,
  log_trans = FALSE,
  chain_col = "chains",
  sep = ";",
  ...
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, cell barcodes should be stored as row names.}

\item{data_cols}{meta.data column(s) containing continuous V(D)J data to
plot}

\item{per_cell}{Should values be plotted per cell, i.e. each data point
would represent one cell. If TRUE, values will be summarized for each cell
using summary_fn. If FALSE, values will be plotted for each chain.}

\item{summary_fn}{Function to use for summarizing values when per_cell is
TRUE, possible values can be either a function, e.g. mean, or a purrr-style
lambda, e.g. ~ mean(.x, na.rm = TRUE) where ".x" refers to the column. If
NULL, the mean will be calculated.}

\item{cluster_col}{meta.data column containing cluster IDs to use for
grouping cells for plotting}

\item{chain}{Chain(s) to use for filtering data before plotting. If NULL
data will not be filtered based on chain.}

\item{type}{Type of plot to create, can be 'histogram', 'density',
'boxplot', or 'violin'}

\item{yaxis}{Units to use for y-axis when type is set to 'histogram'. Use
'frequency' to show number of values or 'percent' to show the percentage of
total values.}

\item{plot_colors}{Character vector specifying colors to use for cell
clusters specified by cluster_col. When cluster_col is NULL, plot colors can
be directly modified with the ggplot2 parameters color and fill,
e.g. fill = "red", color = "black"}

\item{plot_lvls}{Character vector containing order to use for plotting cell
clusters specified by cluster_col}

\item{alpha}{Color transparency}

\item{log_trans}{If TRUE, axis will be log10 transformed}

\item{chain_col}{meta.data column containing chains for each cell}

\item{sep}{Separator used for storing per-chain V(D)J data for each cell}

\item{...}{Additional arguments to pass to ggplot2, e.g. color, fill, size,
linetype, etc.}
}
\value{
ggplot object
}
\description{
Compare V(D)J data for cell clusters
}
\examples{
# Plot the number of indels for each chain
plot_vdj(
  vdj_so,
  data_cols = c("n_insertion", "n_deletion")
)

# Create boxplots
plot_vdj(
  vdj_so,
  data_cols = c("n_insertion", "n_deletion"),
  type = "boxplot"
)

# Pass additional arguments to ggplot2
plot_vdj(
  vdj_sce,
  data_cols = "reads",
  color = "red",
  bins = 25
)

# Compare cell clusters
plot_vdj(
  vdj_sce,
  data_cols = "cdr3_length",
  cluster_col = "orig.ident",
  type = "violin"
)

# log10 transform the axis
plot_vdj(
  vdj_so,
  data_cols = "reads",
  cluster_col = "orig.ident",
  log_trans = TRUE
)

# Express y-axis units as percent of total values
plot_vdj(
  vdj_so,
  data_cols = "n_deletion",
  cluster_col = "orig.ident",
  yaxis = "percent"
)

# Only plot values for heavy chains
plot_vdj(
  vdj_so,
  data_cols = "umis",
  chain = "IGH"
)

# Plot the median number of deletions for each cell
plot_vdj(
  vdj_sce,
  data_cols = "n_deletion",
  per_cell = TRUE,
  summary_fn = stats::median
)

# Set colors for cell clusters
plot_vdj(
  vdj_so,
  data_cols = "cdr3_length",
  cluster_col = "orig.ident",
  plot_colors = c(avid_1 = "red", avid_2 = "purple")
)

# Set order to use for plotting cell clusters
plot_vdj(
  vdj_so,
  data_cols = "cdr3_length",
  cluster_col = "orig.ident",
  plot_lvls = c("avid_2", "avid_1"),
  type = "boxplot"
)

}
\seealso{
\code{\link[=summarize_vdj]{summarize_vdj()}} for more examples on how per-chain data can be
summarized for each cell
}
