% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{splen_so}
\alias{splen_so}
\title{A small Seurat object containing downsampled splenocyte data}
\format{
A Seurat object with 10701 features, 500 cells, and 1 assay
}
\usage{
splen_so
}
\description{
A small Seurat object containing downsampled splenocyte data
}
\keyword{datasets}
