% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-vdj.R
\name{summarize_vdj}
\alias{summarize_vdj}
\title{Summarize V(D)J data for each cell}
\usage{
summarize_vdj(
  input,
  vdj_cols,
  fn = NULL,
  ...,
  chain = NULL,
  chain_col = "chains",
  sep = ";",
  col_names = "{.col}",
  return_df = FALSE
)
}
\arguments{
\item{input}{Single cell object or data.frame containing V(D)J data. If a
data.frame is provided, the cell barcodes should be stored as row names.}

\item{vdj_cols}{meta.data column(s) containing V(D)J data to summarize for
each cell}

\item{fn}{Function to apply to each selected column, possible values can be
either a function, e.g. mean, or a purrr-style lambda, e.g. ~ mean(.x,
na.rm = TRUE). If NULL, the mean will be calculated for numeric values,
non-numeric columns will be combined into a single string.}

\item{...}{Additional arguments to pass to fn}

\item{chain}{Chain to use for summarizing V(D)J data}

\item{chain_col}{meta.data column(s) containing chains for each cell}

\item{sep}{Separator used for storing per cell V(D)J data}

\item{col_names}{A glue specification that describes how to name the output
columns, this can use {.col} to stand for the selected column name}

\item{return_df}{Return results as a data.frame. If FALSE, results will be
added to the input object.}
}
\value{
data.frame containing V(D)J data summarized for each cell
}
\description{
Summarize values for each cell using a function or purrr-style lambda. This
is useful since some V(D)J metrics include data for each chain. Per-chain
data will be summarized, which is helpful for plotting these metrics per
cell.
}
