% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_so}
\alias{tiny_so}
\title{Seurat object for tests}
\format{
A Seurat object with 200 features, 200 cells, and 1 assay
}
\usage{
tiny_so
}
\description{
A tiny Seurat object for tests
}
\keyword{datasets}
