% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_vdj}
\alias{tiny_vdj}
\title{Seurat object with V(D)J data for tests}
\format{
A Seurat object with 200 features, 200 cells, and 1 assay
}
\usage{
tiny_vdj
}
\description{
A tiny Seurat object with V(D)J data for tests
}
\keyword{datasets}
