% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vdj_theme}
\alias{vdj_theme}
\title{Theme for djvdj plotting functions}
\usage{
vdj_theme(txt_size = 11, ttl_size = 12, txt_col = "black")
}
\arguments{
\item{txt_size}{Size of axis text}

\item{ttl_size}{Size of axis titles}

\item{txt_col}{Color of axis text}
}
\value{
ggplot theme
}
\description{
Theme for djvdj plotting functions
}
